% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-downAndUpload.R
\name{addFilesFromOtherZip}
\alias{addFilesFromOtherZip}
\title{Add Internal Zip File}
\usage{
addFilesFromOtherZip(tempDir, pathToOtherZip)
}
\arguments{
\item{tempDir}{(character) temporary directory to store the help file}

\item{pathToOtherZip}{(reactive) path to another zip file. The content will be added to the
zip file that contains the content to download. Needed in MapR. If NULL, no files are added.}
}
\description{
Add Internal Zip File
}
