% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-downAndUpload.R
\name{addHelpFile}
\alias{addHelpFile}
\title{Add Help File}
\usage{
addHelpFile(tempDir, helpHTML)
}
\arguments{
\item{tempDir}{(character) temporary directory to store the help file}

\item{helpHTML}{content of help function}
}
\description{
Add Help File
}
