% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-downAndUpload.R
\name{addInternalZipFile}
\alias{addInternalZipFile}
\title{Add Internal Zip File}
\usage{
addInternalZipFile(filesToZip, pathToInnerZip)
}
\arguments{
\item{filesToZip}{(character) files to be zipped}

\item{pathToInnerZip}{(character) path to inner zip file}
}
\value{
(character) updated paths to files to zip
}
\description{
Add Internal Zip File
}
