% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-downAndUpload.R
\name{addModelRDSFile}
\alias{addModelRDSFile}
\title{Add Model RDS File}
\usage{
addModelRDSFile(
  tempDir,
  dat,
  inputs,
  model,
  rPackageName,
  subFolder,
  onlySettings
)
}
\arguments{
\item{tempDir}{(character) temporary directory to store the model file}

\item{dat}{(reactive) user data}

\item{inputs}{(reactiveValues) reactiveValues list of user inputs, in most cases just the
"inputs" list}

\item{model}{(reactive) model output object}

\item{rPackageName}{(character) name of the package (as in the Description file) in which this
module is called. If not NULL, than the uploaded file must be a downloaded file
from the R package where \code{importDataServer} was called. This parameter is ignored if
\code{importType == "data"}.}

\item{subFolder}{(character) (optional) subfolder containing loadable .zip files}

\item{onlySettings}{(logical) if TRUE allow only download of user inputs and user data}
}
\description{
Add Model RDS File
}
