% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-downAndUpload.R
\name{addNotesFile}
\alias{addNotesFile}
\title{Add Notes File}
\usage{
addNotesFile(tempDir, notes)
}
\arguments{
\item{tempDir}{(character) temporary directory to store the notes file}

\item{notes}{(character) notes to be added to the zip file}
}
\description{
Add Notes File
}
