% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-dataSource.R
\name{addSourceType}
\alias{addSourceType}
\title{Get Source Type}
\usage{
addSourceType(dataSource, importType, source, inputDataOrLink)
}
\arguments{
\item{dataSource}{(reactiveValues) path, filename, type and input, output of \code{selectSourceServer()}}

\item{importType}{(character) what kind of import is expected: either "data", "model" or "zip"}

\item{source}{(character) selected option to load data from: either "ckan", "file", "url" or "remoteModel"}

\item{inputDataOrLink}{(character) either "fullData" or "dataLink" option}
}
\description{
Get Source Type
}
