% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-DataProcessItem-class.R
\name{as.DataProcessLink.DataProcessItem}
\alias{as.DataProcessLink.DataProcessItem}
\title{S3 method: Convert DataProcessItem to DataProcessLink
Converts a DataProcessItem object to a DataProcessLink object.}
\usage{
\method{as.DataProcessLink}{DataProcessItem}(object, ...)
}
\arguments{
\item{object}{The DataProcessItem object to convert.}

\item{...}{Additional arguments (not used).}
}
\value{
A DataProcessLink object.
}
\description{
S3 method: Convert DataProcessItem to DataProcessLink
Converts a DataProcessItem object to a DataProcessLink object.
}
