% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-DataProcessLink-class.R
\name{as.list.DataProcessLink}
\alias{as.list.DataProcessLink}
\title{S3 method: Convert DataProcessLink to a plain list
Converts a DataProcessLink object to a plain list, excluding class attributes.}
\usage{
\method{as.list}{DataProcessLink}(x, ...)
}
\arguments{
\item{x}{The DataProcessLink object to convert.}

\item{...}{Additional arguments (not used).}
}
\value{
A plain list representation of the DataProcessLink x.
}
\description{
S3 method: Convert DataProcessLink to a plain list
Converts a DataProcessLink object to a plain list, excluding class attributes.
}
