% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-DownloadBundle-class.R
\name{buildDownloadZip}
\alias{buildDownloadZip}
\title{Build a download bundle zip file}
\usage{
buildDownloadZip(
  out_file = tempfile(fileext = ".zip"),
  rPackageName = NULL,
  dat = NULL,
  inputs = NULL,
  model = NULL,
  sub_model = NULL,
  helpHTML = NULL,
  notes = NULL,
  onlySettings = FALSE,
  otherPaths = NULL,
  otherRoot = NULL,
  compressionLevel = 9
)
}
\arguments{
\item{out_file}{The path to the output zip file.}

\item{rPackageName}{The name of the R package for which the bundle is created.}

\item{dat}{The model data to be included in the bundle.}

\item{inputs}{The model inputs to be included in the bundle.}

\item{model}{The model object to be included in the bundle.}

\item{sub_model}{Optional sub-model name within the R package.}

\item{helpHTML}{Optional HTML content for help documentation.}

\item{notes}{Optional notes to be included in the bundle.}

\item{onlySettings}{Logical indicating whether to include only settings (inputs) in the bundle.}

\item{otherPaths}{Optional character vector of additional file/directory paths to include in the
bundle.}

\item{otherRoot}{Optional root directory for the additional paths to preserve relative structure.}

\item{compressionLevel}{The level of compression to use when zipping the bundle (default is 9).}
}
\value{
The path to the created zip file (invisible).
}
\description{
This function creates a zip file containing the specified model data, inputs,
notes, and help documentation. It utilizes the DownloadBundle class to manage
the bundling process.
}
