% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-isomemoAPI.R
\name{callAPI}
\alias{callAPI}
\title{Call API}
\usage{
callAPI(action, ...)
}
\arguments{
\item{action}{(character) name of the endpoint, one of "mapping-ids", "dbsources", "iso-data" or
"mapping"}

\item{...}{parameters for the endpoint, e.g. mappingId = "IsoMemo", dbsource = "LiVES,
field = "site,longitude", ...}
}
\description{
Call API
}
