% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-checks.R
\name{checkAnyNonNumericColumns}
\alias{checkAnyNonNumericColumns}
\title{Check Any Non-Numeric Columns}
\usage{
checkAnyNonNumericColumns(data)
}
\arguments{
\item{data}{data to be checked}
}
\description{
Check for any non-numeric columns. Check can be applied as custom checks in importDataServer.
}
