% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-checks.R
\name{checkErrorAnyNonNumericColumns}
\alias{checkErrorAnyNonNumericColumns}
\title{Check Error Any Non-Numeric Columns}
\usage{
checkErrorAnyNonNumericColumns(data)
}
\arguments{
\item{data}{data to be checked}
}
\description{
Check for any non-numeric columns. Check can be applied as customErrorChecks in importDataServer.
}
