% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-checks.R
\name{checkErrorNoNumericColumns}
\alias{checkErrorNoNumericColumns}
\title{Check Error No Numeric Columns}
\usage{
checkErrorNoNumericColumns(data)
}
\arguments{
\item{data}{data to be checked}
}
\description{
Check for minimal numeric columns. Check can be applied as customErrorChecks in importDataServer.
}
