% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-loadImport.R
\name{checkExtension}
\alias{checkExtension}
\title{Check Extension}
\usage{
checkExtension(filepath, fileExtension = "zip", defaultExtension = "zip")
}
\arguments{
\item{filepath}{(character) path to the file}

\item{fileExtension}{(character) expected file extension}

\item{defaultExtension}{(character) default file extension}
}
\description{
Check if the file extension of a file is valid.
}
