% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-remoteModels.R
\name{checkLocalModelDir}
\alias{checkLocalModelDir}
\title{Check Local Model Dir}
\usage{
checkLocalModelDir(pathToLocal)
}
\arguments{
\item{pathToLocal}{(character) relative path to the folder storing local files}
}
\description{
Check Local Model Dir
}
