% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-checks.R
\name{checkNonNumericColumnsExceptFirst}
\alias{checkNonNumericColumnsExceptFirst}
\title{Check Non-Numeric Columns Except First}
\usage{
checkNonNumericColumnsExceptFirst(data)
}
\arguments{
\item{data}{data to be checked}
}
\description{
Check for any non-numeric columns, exclude the first column. Check can be applied as custom
 checks in importDataServer.
}
