% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-checks.R
\name{checkWarningEmptyValues}
\alias{checkWarningEmptyValues}
\title{Check Warning Empty VAlues}
\usage{
checkWarningEmptyValues(data)
}
\arguments{
\item{data}{data to be checked}
}
\description{
Check for empty values that can be applied as customWarningChecks in importDataServer.
}
