% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-importData-configureData.R
\name{configureDataServer}
\alias{configureDataServer}
\title{Configure Data Server}
\usage{
configureDataServer(
  id,
  mergeList,
  customNames,
  dataSource,
  ignoreWarnings = FALSE
)
}
\arguments{
\item{id}{id of module}

\item{mergeList}{(reactiveVal) list of data imports submitted for data processing via buttons
'Create Query with data' or 'Prepare / Merge data'}

\item{customNames}{settings for custom column and row names}

\item{dataSource}{(reactiveValues) path, filename, type and input, output of \code{selectSourceServer()}}

\item{ignoreWarnings}{(logical) TRUE to enable imports in case of warnings}
}
\description{
Server function of the module
}
