% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-importModule-configureFile.R
\name{configureFileUI}
\alias{configureFileUI}
\title{Configure File UI}
\usage{
configureFileUI(
  id,
  customHelpText = importOptions()[["customHelpText"]],
  defaultFileTypes = config()[["dataFileTypes"]],
  userFileTypes = c()
)
}
\arguments{
\item{id}{id of module}

\item{customHelpText}{(list) A help text element that can be added to a UI definition. Output of
\code{shiny::helpText(...)}.}

\item{defaultFileTypes}{(character) default file types}

\item{userFileTypes}{(character) user file types specified in "Pandora Platform" settings}
}
\description{
UI of the module
}
