% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-importData.R
\name{dataImportOptions}
\alias{dataImportOptions}
\title{Data Import Options}
\usage{
dataImportOptions(
  ckanFileTypes = c("xls", "xlsx", "csv", "odt", "txt"),
  rowNames = reactiveVal(NULL),
  colNames = reactiveVal(NULL),
  customWarningChecks = list(),
  customErrorChecks = list(),
  batch = FALSE,
  outputAsMatrix = FALSE
)
}
\arguments{
\item{ckanFileTypes}{(character) file types allowed for import from Pandora ("ckan")}

\item{rowNames}{(reactive) use this for rownames of imported data. This parameter is ignored if importType == "model"}

\item{colNames}{(reactive) use this for colnames of imported data. This parameter is ignored if importType == "model"}

\item{customWarningChecks}{list of reactive(!) functions which will be executed after importing
of data.
 functions need to return TRUE if check is successful or a character with a warning otherwise.
 This parameter is ignored if importType == "model"}

\item{customErrorChecks}{list of reactive(!) functions which will be executed after importing
of data.
  functions need to return TRUE if check is successful or a character with a warning otherwise.
  This parameter is ignored if importType == "model"}

\item{batch}{(logical) use batch import. This parameter is ignored if importType == "model"}

\item{outputAsMatrix}{(logical) TRUE if output must be a matrix,
e.g. for batch = TRUE in Resources. This parameter is ignored if importType == "model"}
}
\value{
list of extra options for \code{impoertDataServer()} that are relevant if
\code{importType = "data"}
}
\description{
Data Import Options
}
