% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-dataModule.R
\name{dataUI}
\alias{dataUI}
\alias{dataServer}
\title{UI function of data module}
\usage{
dataUI(id, title = "Data", titleTag = "h4")

dataServer(id, path = NULL, transformations = list(), ...)
}
\arguments{
\item{id}{module id}

\item{title}{(character) module title}

\item{titleTag}{(character) HTML tag to put around the title, e.g. "h4" for \code{h4} from
\code{htmltools}}

\item{path}{path to the example data file, e.g. \code{file.path("data", "example.csv")}}

\item{transformations}{list of transformations to apply to the dataset}

\item{...}{further arguments passed to \code{DataTools::importDataServer}}
}
\description{
Wrapper module for the \code{DataTools::importDataServer} module that allows for loading example data
}
