% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-importData-prepareData.R
\name{deleteColumnsUI}
\alias{deleteColumnsUI}
\alias{deleteColumnsServer}
\title{Delete Columns UI}
\usage{
deleteColumnsUI(id)

deleteColumnsServer(id, preparedData)
}
\arguments{
\item{id}{id of module}

\item{preparedData}{(reactive) selected data, possibly already modified}
}
\description{
UI of the module

Server function of the module
}
