% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-remoteModels.R
\name{downloadFileToTmp}
\alias{downloadFileToTmp}
\title{Download Zip To Tmp}
\usage{
downloadFileToTmp(url, fileext = ".zip")
}
\arguments{
\item{url}{a \code{\link{character}} string (or longer vector
    for the \code{"libcurl"} method) naming the URL of a resource to be
    downloaded.}

\item{fileext}{a non-empty character vector giving the file extension}
}
\value{
(character) path to the temporary zip file
}
\description{
Download Zip To Tmp
}
