% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-downAndUpload.R
\name{downloadModelServer}
\alias{downloadModelServer}
\title{Server function download model}
\usage{
downloadModelServer(
  id,
  dat,
  inputs,
  model,
  customList = reactive(list()),
  rPackageName,
  subFolder = NULL,
  fileExtension = "zip",
  helpHTML = "",
  modelNotes = reactive(""),
  triggerUpdate = reactive(TRUE),
  onlySettings = FALSE,
  compress = TRUE,
  compressionLevel = 9
)
}
\arguments{
\item{id}{namespace id}

\item{dat}{(reactive) user data}

\item{inputs}{(reactiveValues) reactiveValues list of user inputs, in most cases just the
default "inputs" list from shiny}

\item{model}{(reactive) model output object}

\item{customList}{(reactive) reactive list of objects to be stored for later upload, e.g. a list
of lists that store output of modules for user inputs}

\item{rPackageName}{(character) name of the package (as in the Description file) in which this
module is called. If not NULL, than the uploaded file must be a downloaded file
from the R package where \code{importDataServer} was called. This parameter is ignored if
\code{importType == "data"}.}

\item{subFolder}{(character) (optional) subfolder containing loadable .zip files}

\item{fileExtension}{(character) (optional) app specific file extension, e.g. "resources",
"bpred", "bmsc"}

\item{helpHTML}{content of help function}

\item{modelNotes}{(reactive) notes regarding the object to be saved and displayed when uploaded}

\item{triggerUpdate}{(reactive) trigger the update of the "Notes" text input. Useful, when
applying this module inside a modal window.}

\item{onlySettings}{(logical) if TRUE allow only download of user inputs and user data}

\item{compress}{a logical specifying whether saving to a named file is to use "gzip" compression,
or one of "gzip", "bzip2" or "xz" to indicate the type of compression to be used. Ignored if
file is a connection.}

\item{compressionLevel}{A number between 1 and 9. 9 compresses best, but it also takes the
longest.}
}
\description{
Backend for download model module
}
