% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-downAndUpload.R
\name{downloadModelUI}
\alias{downloadModelUI}
\title{Download model module}
\usage{
downloadModelUI(id, label, width = NULL)
}
\arguments{
\item{id}{id of module}

\item{label}{title of module}

\item{width}{width of inputs in percent}
}
\description{
UI function to download a zip file with notes and a list of models
}
