% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-downAndUpload.R
\name{downUploadButtonUI}
\alias{downUploadButtonUI}
\title{downUploadButton module ui}
\usage{
downUploadButtonUI(id, title = NULL, label = "Download / Upload Model")
}
\arguments{
\item{id}{module id}

\item{title}{title of the module}

\item{label}{label for actionButton which will open a modal}
}
\description{
A button that opens a modal containing the UI to download or to upload model objects
}
