% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-remoteModels.R
\name{exampleOptions}
\alias{exampleOptions}
\title{Extra options for remote examples}
\usage{
exampleOptions(
  pathToLocal = file.path(".", "predefinedModels"),
  folderOnGithub = "/predefinedModels",
  fileExtension = "zip"
)
}
\arguments{
\item{pathToLocal}{(character) relative path to the folder storing local files}

\item{folderOnGithub}{(character) folder on github where remote files are stored. This should
correspond to 'pathToLocal' since online and offline files should be the same and up-to-date}

\item{fileExtension}{(character) (otional) app specific file extension, e.g. "resources", "bmsc",
"bpred", or (app-unspecific) "zip". Only files with this extension are valid for import.}
}
\value{
(list) the path to the selected remote (github) or local file
}
\description{
Extra options for \code{remoteModelsServer()}
}
