% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-importBundle-class.R
\name{extract_model_import}
\alias{extract_model_import}
\title{Extract model import from bundle import
Extracts either full model or inputs-only from a bundle import.}
\usage{
extract_model_import(bundle_import)
}
\arguments{
\item{bundle_import}{Result of import_bundle_zip().}
}
\value{
Model import (list) or NULL if not found.
}
\description{
Extract model import from bundle import
Extracts either full model or inputs-only from a bundle import.
}
