% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-importBundle-class.R
\name{extract_model_notes}
\alias{extract_model_notes}
\title{Extract model notes from bundle import
Extracts README.txt contents from a bundle import.}
\usage{
extract_model_notes(bundle_import)
}
\arguments{
\item{bundle_import}{Result of import_bundle_zip().}
}
\value{
Model notes (character) or NULL if not found.
}
\description{
Extract model notes from bundle import
Extracts README.txt contents from a bundle import.
}
