% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-DataProcessItem-class.R
\name{extract_unique_inputs.DataProcessItem}
\alias{extract_unique_inputs.DataProcessItem}
\title{S3 method: Extract unique inputs for DataProcessItem
Extracts unique user inputs from a DataProcessItem object.}
\usage{
\method{extract_unique_inputs}{DataProcessItem}(object, ...)
}
\arguments{
\item{object}{The DataProcessItem object to extract from.}

\item{...}{Additional arguments (not used).}
}
\value{
A named list of unique user inputs.
}
\description{
S3 method: Extract unique inputs for DataProcessItem
Extracts unique user inputs from a DataProcessItem object.
}
