% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-DataProcessLink-class.R
\name{extract_unique_inputs.DataProcessLink}
\alias{extract_unique_inputs.DataProcessLink}
\title{S3 method: Extract unique inputs for DataProcessLink
Extracts unique user inputs from a DataProcessLink object.}
\usage{
\method{extract_unique_inputs}{DataProcessLink}(object, ...)
}
\arguments{
\item{object}{The DataProcessLink object to extract from.}

\item{...}{Additional arguments (not used).}
}
\value{
A named list of unique user inputs.
}
\description{
S3 method: Extract unique inputs for DataProcessLink
Extracts unique user inputs from a DataProcessLink object.
}
