% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-loadModel.R
\name{extractInputsFromModel}
\alias{extractInputsFromModel}
\title{Extract Inputs From Model}
\usage{
extractInputsFromModel(modelImport, rPackageName)
}
\arguments{
\item{modelImport}{(list) model object}

\item{rPackageName}{(character) name of the app}
}
\value{
(list) inputs object
}
\description{
Dependent on the app, the inputs object is stored in different places. This function extracts the
inputs object from the model object.
}
