% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-loadModel.R
\name{extractModelFromFile}
\alias{extractModelFromFile}
\title{Extract Model From File}
\usage{
extractModelFromFile(pathToUnzipped)
}
\arguments{
\item{pathToUnzipped}{(character) path to the folder were the model was unzipped}
}
\description{
Extracts the model object from either a "model.rds" file or from a "model.Rdata" file.
Recent model objects are only stored as "model.rds".
}
