% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-loadModel.R
\name{extractModelFromModel}
\alias{extractModelFromModel}
\title{Extract Model From Model}
\usage{
extractModelFromModel(modelImport, rPackageName = NULL)
}
\arguments{
\item{modelImport}{(list) model object}

\item{rPackageName}{(character) name of the app}
}
\value{
(list) model object
}
\description{
Extract Model From Model
}
