% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-loadModel.R
\name{extractNotes}
\alias{extractNotes}
\title{Extract Notes}
\usage{
extractNotes(pathToUnzipped)
}
\arguments{
\item{pathToUnzipped}{(character) path to the folder were the model (or inputs) was unzipped.}
}
\value{
(character) notes
}
\description{
Extracts the notes from the README.txt file.
}
