% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-importData-mergeData.R
\name{extractTableIds}
\alias{extractTableIds}
\title{Extract Table IDs}
\usage{
extractTableIds(mergeList)
}
\arguments{
\item{mergeList}{(list) list of data to be merged}
}
\value{
(character) short internal table names
}
\description{
Create IDs `1:n` to the names of loaded tables. Names are often url's of the table file,
something like "https://pandoradata.earth/dataset/.../download/afriarch-isotopic-dataset.xlsx".
}
