% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-isomemoAPI.R
\name{fillIsoData}
\alias{fillIsoData}
\title{Fill Iso Data}
\usage{
fillIsoData(data, mapping)
}
\arguments{
\item{data}{(data.frame) Output from IsoMemo::getData(). A data frame containing the
requested databases, category domains, and variables of interest from the user}

\item{mapping}{(data.frame) Output from IsoMemo::getFields() A data frame that describes data
field name, data type, and domain category}
}
\value{
A data frame also containing the columns defined in the mapping but missing in the data.
 New columns have all values set to NA.
}
\description{
Fill Iso Data
}
