% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-importData-loadImport.R
\name{fillValuesFromModel}
\alias{fillValuesFromModel}
\title{Fill Values From Model}
\usage{
fillValuesFromModel(values, filename, importedModel)
}
\arguments{
\item{values}{(reactiveValues) empty list of values in the format of the output of loadDataWrapper}

\item{filename}{(reactive) name of the loaded file}

\item{importedModel}{(list) output of loadModel()}
}
\value{
(list) list of values in the format of the output of loadDataWrapper
}
\description{
Format list of model data to be compatible with the import module such that success, warnings and
 errors are displayed inside the pop-up modal.
}
