% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-loadImport.R
\name{fillValuesFromZip}
\alias{fillValuesFromZip}
\title{Fill Values From Zip}
\usage{
fillValuesFromZip(values, filename, importZip, expectedFileInZip)
}
\arguments{
\item{values}{(reactiveValues) empty list of values in the format of the output of loadDataWrapper}

\item{filename}{(character) name of the loaded file}

\item{importZip}{(list) output of loadModel()}

\item{expectedFileInZip}{(character) (optional) This parameter is ignored if importType != "zip".
File names that must be contained in the zip upload.}
}
\value{
(list) list of values in the format of the output of loadDataWrapper
}
\description{
Set values, warning and error messages for the import module such that success, warnings and
 errors are displayed inside the pop-up modal.
}
