% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-ckan.R
\name{filterCKANByMeta}
\alias{filterCKANByMeta}
\title{Filter CKAN by Meta}
\usage{
filterCKANByMeta(fileList, meta = "")
}
\arguments{
\item{fileList}{(list) output from the Pandora API}

\item{meta}{(character) string for filtering all meta information}
}
\value{
(list) a fileList where the entries meta data contains the string 'meta'
}
\description{
Filter CKAN by Meta
}
