% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-ckan.R
\name{filterCKANGroup}
\alias{filterCKANGroup}
\title{Filter CKAN by Group}
\usage{
filterCKANGroup(fileList, ckanGroup = NA)
}
\arguments{
\item{fileList}{(list) output from the Pandora API}

\item{ckanGroup}{(character) title of a CKAN group}
}
\value{
(list) a fileList where the entries 'groups' == ckanGroup
}
\description{
Filter CKAN by Group
}
