% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-ckan.R
\name{filterSingleCKANRecord}
\alias{filterSingleCKANRecord}
\title{Filter Single CKAN Record}
\usage{
filterSingleCKANRecord(record)
}
\arguments{
\item{record}{(list) single entry from fileList}
}
\description{
Removes all meta information that is not needed by selecting only relevant entries.
}
