% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-importData-linkToData.R
\name{filterUnprocessed}
\alias{filterUnprocessed}
\title{Filter Unprocessed}
\usage{
filterUnprocessed(mergeList)
}
\arguments{
\item{mergeList}{(reactiveVal) list of data imports submitted for data processing via buttons
'Create Query with data' or 'Prepare / Merge data'}
}
\description{
Filter Unprocessed
}
