% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-importData.R
\name{formatColumnNames}
\alias{formatColumnNames}
\title{Format Column Names}
\usage{
formatColumnNames(vNames, silent = FALSE)
}
\arguments{
\item{vNames}{(character) names of the imported data's columns}

\item{silent}{(logical) set TRUE prevent notification of warnings}
}
\description{
Replaces all not alpha-numeric characters in the names of columns with a dot.
}
