% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-importData-selectFileType.R
\name{formatTypeChoices}
\alias{formatTypeChoices}
\title{Format Type Choices}
\usage{
formatTypeChoices(typeChoices)
}
\arguments{
\item{typeChoices}{(character) vector of file types}
}
\description{
Format Type Choices
}
