% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-ckan.R
\name{getCKANRecordChoices}
\alias{getCKANRecordChoices}
\title{Get CKAN Record Choices}
\usage{
getCKANRecordChoices(network = "", pattern = "")
}
\arguments{
\item{network}{(character) name of Pandora network}

\item{pattern}{(character) string for meta information search}
}
\description{
Get choices that will be available in the ckanRecord (repository) input
}
