% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-ckan.R
\name{getCKANResourcesChoices}
\alias{getCKANResourcesChoices}
\title{Get CKAN Resource Choices}
\usage{
getCKANResourcesChoices(
  fileType = character(),
  repository = "",
  network = "",
  pattern = "",
  packageList = data.frame()
)
}
\arguments{
\item{fileType}{(character) list of relevant file types, e.g. c("xls", "xlsx", "csv", "odt")}

\item{repository}{(character) name of a Pandora repository, e.g. an entry of the output from
\code{getRepositories()$name}}

\item{network}{(character) name of a Pandora network, e.g. an entry of the output from
\code{getNetworks()$name}}

\item{pattern}{(character) string for meta information search}

\item{packageList}{(data.frame) optional, output of callAPI() from a previous call to the
Pandora API.}
}
\description{
Get choices that will be available in the ckanResource input
}
