% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-isomemoAPI.R
\name{getDatabaseList}
\alias{getDatabaseList}
\title{Get Database List}
\usage{
getDatabaseList(mappingId = "IsoMemo")
}
\arguments{
\item{mappingId}{(character) If desired, provide a different mappingId in order to obtain a list
of databases for that mapping. Check available mapping ids with getMappingIds().}
}
\description{
Get Database List
}
