% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-importData-linkToData.R
\name{getFileInputs}
\alias{getFileInputs}
\title{Get File Inputs}
\usage{
getFileInputs(input, type = c("file", "source"))
}
\arguments{
\item{input}{(reactiveValue) input}

\item{type}{(character) type of inputs}
}
\description{
Filter all inputs for file inputs or for source inputs
}
