% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-downAndUpload.R
\name{getFolderOnGithub}
\alias{getFolderOnGithub}
\title{Get Folder on Github}
\usage{
getFolderOnGithub(mainFolder, subFolder = NULL)
}
\arguments{
\item{mainFolder}{(character) folder containing all loadable .zip files. For most apps this
is folder "predefinedModels". In most apps it can be found under "inst/app/".}

\item{subFolder}{(character) (optional) subfolder containing loadable .zip files}
}
\description{
Get Folder on Github
}
