% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-remoteModels.R
\name{getGithubContent}
\alias{getGithubContent}
\title{Get Github Content}
\usage{
getGithubContent(githubRepo)
}
\arguments{
\item{githubRepo}{(character) name of used github repository, e.g. "bpred"}
}
\description{
Get content of api call to github folder
}
