% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-remoteModels.R
\name{getLocalModelDir}
\alias{getLocalModelDir}
\title{Get Local Model Dir}
\usage{
getLocalModelDir(githubRepo, customPathToLocalModels = NULL)
}
\arguments{
\item{githubRepo}{(character) name of used github repository, e.g. "bpred"}

\item{customPathToLocalModels}{(character) custom relative path to the folder storing local models.
If assuming that the package 'DataTools' was installed into the same parent directory as the
the app's package in which this module should be applied, than the path for e.g. the package
"mpiBpred" would "../bpred/inst/app/predefinedModels"}
}
\description{
Get Local Model Dir
}
