% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-remoteModels.R
\name{getLocalModels}
\alias{getLocalModels}
\title{Get Local Models}
\usage{
getLocalModels(pathToLocal, fileExtension = "zip")
}
\arguments{
\item{pathToLocal}{(character) relative path to the folder storing local files}

\item{fileExtension}{(character) (otional) app specific file extension, e.g. "resources", "bmsc",
"bpred", or (app-unspecific) "zip". Only files with this extension are valid for import.}
}
\description{
Get Local Models
}
