% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-isomemoAPI.R
\name{getMappingTable}
\alias{getMappingTable}
\title{getMappingTable}
\usage{
getMappingTable(mappingId = "IsoMemo")
}
\arguments{
\item{mappingId}{(character) If desired, provide a different mappingId in order to obtain the
mapping table for that mapping. Check available mapping ids with getMappingIds().}
}
\description{
getMappingTable
}
