% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-isomemoAPI.R
\name{getRemoteData}
\alias{getRemoteData}
\title{getRemoteData}
\usage{
getRemoteData(db, mappingId = "IsoMemo")
}
\arguments{
\item{db}{database}

\item{mappingId}{(character) If desired, provide a different mappingId in order to obtain the
data for that mapping. Check available mapping ids with getMappingIds().}
}
\description{
getRemoteData
}
