% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-modelUpload-remote.R
\name{getRemoteModelsFromGithub}
\alias{getRemoteModelsFromGithub}
\title{Get Remote Models From Github}
\usage{
getRemoteModelsFromGithub(githubRepo)
}
\arguments{
\item{githubRepo}{(character) name of the app's github repository, currently "bpred",
"osteo-bior", "resources"}
}
\description{
Get remote models from github directory
}
