% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-dataSource.R
\name{getSourceFile}
\alias{getSourceFile}
\title{Get Source File}
\usage{
getSourceFile(dataSource, input)
}
\arguments{
\item{dataSource}{(reactiveValues)}

\item{input}{(reactiveValues)}
}
\description{
Get Source File
}
