% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-dataSource.R
\name{getSourceModel}
\alias{getSourceModel}
\title{Get Source model}
\usage{
getSourceModel(dataSource, input, pathToFile = NULL)
}
\arguments{
\item{dataSource}{(reactiveValues)}

\item{input}{(reactiveValues)}

\item{pathToFile}{(character) file path, ignored if \code{type != "remoteModel"}}
}
\description{
Get Source model
}
