% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-loadModel.R
\name{getZip}
\alias{getZip}
\title{Get Zip}
\usage{
getZip(filepath)
}
\arguments{
\item{filepath}{(character) path to the model file}
}
\value{
(character) path to the model file
}
\description{
Get Zip
}
