% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-rgpt3.R
\name{gpt3_test_completion}
\alias{gpt3_test_completion}
\title{Make a test request to the GPT-3 API}
\usage{
gpt3_test_completion(verbose = TRUE)
}
\arguments{
\item{verbose}{(boolean) if TRUE prints the actual prompt and GPT-3
completion of the test request (default: TRUE).}
}
\value{
A message of success or failure of the connection.

 This function was copied from https://github.com/ben-aaron188/rgpt3.
}
\description{
`gpt3_test_completion()` sends a basic [completion
  request](https://beta.openai.com/docs/api-reference/completions) to the
  Open AI GPT-3 API.
}
