% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-remoteModels.R
\name{has_internet}
\alias{has_internet}
\title{Has Internet}
\usage{
has_internet(timeout = 2)
}
\arguments{
\item{timeout}{(numeric) number of seconds to wait for a response until giving up. Can not be less than 1 ms.}
}
\description{
Has Internet
}
