% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-importBundle-class.R
\name{import_bundle_zip}
\alias{import_bundle_zip}
\title{Convenience: import zip, index it, optionally load known parts}
\usage{
import_bundle_zip(
  zipfile,
  extract_dir = NULL,
  keep_dir = FALSE,
  include_hidden = FALSE,
  load_known = TRUE,
  from_dir = FALSE
)
}
\arguments{
\item{zipfile}{Path to zip.}

\item{extract_dir}{Optional directory to extract to (NULL = temp).}

\item{keep_dir}{Keep extracted files if temp dir is used.}

\item{include_hidden}{Include dotfiles.}

\item{load_known}{Logical; if TRUE returns loaded known parts as well.}

\item{from_dir}{Logical; if TRUE, treat 'zipfile' as a directory path instead of a zip file.
Assumes files are already extracted.}
}
\value{
A list with $zip_import (ZipImport), $available (list), and optionally $loaded.
}
\description{
Convenience: import zip, index it, optionally load known parts
}
