% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-importData.R
\name{importDataServer}
\alias{importDataServer}
\title{Server function for data import}
\usage{
importDataServer(
  id,
  rowNames = reactiveVal(NULL),
  colNames = reactiveVal(NULL),
  customWarningChecks = list(),
  customErrorChecks = list(),
  ignoreWarnings = FALSE,
  defaultSource = "ckan",
  batch = FALSE,
  outputAsMatrix = FALSE
)
}
\arguments{
\item{id}{namespace id}

\item{rowNames}{(reactive) use this for rownames of imported data}

\item{colNames}{(reactive) use this for colnames of imported data}

\item{customWarningChecks}{list of reactive(!) functions which will be executed after importing
of data.
 functions need to return TRUE if check is successful or a character with a warning otherwise.}

\item{customErrorChecks}{list of reactive(!) functions which will be executed after importing
of data.
  functions need to return TRUE if check is successful or a character with a warning otherwise.}

\item{ignoreWarnings}{TRUE to enable imports in case of warnings}

\item{defaultSource}{(character) default source for input "Source", e.g. "ckan", "file", or "url"}

\item{batch}{(logical) use batch import}

\item{outputAsMatrix}{(logical) TRUE if output must be a matrix,
e.g. for batch = TRUE in Resources}
}
\description{
Backend for data import module
}
