% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-importModule-configureFile.R
\name{importMessageServer}
\alias{importMessageServer}
\title{Import Message Server}
\usage{
importMessageServer(id, values)
}
\arguments{
\item{id}{id of module}

\item{values}{(reactiveValues) values of the import module}
}
\description{
Server function of the module
}
