% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-importOptions.R
\name{importOptions}
\alias{importOptions}
\title{Import Options}
\usage{
importOptions(rPackageName = "", githubRepo = "", customHelpText = NULL)
}
\arguments{
\item{rPackageName}{(character) name of the package (as in the Description file) in which this
module is called. If not NULL, than the uploaded file must be a downloaded file
from the R package where \code{importDataServer} was called. This parameter is ignored if
\code{importType == "data"}.}

\item{githubRepo}{(character) name of used github repository, e.g. "bpred"}

\item{customHelpText}{(list) A help text element that can be added to a UI definition. Output of
\code{shiny::helpText(...)}.}
}
\description{
Extra options for the import module.
}
