% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-updateListNamesIfDuplicate.R
\name{incIndexOfName}
\alias{incIndexOfName}
\title{Inc Index Of Name}
\usage{
incIndexOfName(name)
}
\arguments{
\item{name}{(character) name}
}
\description{
If the name has no index, add a new index: "(1)". If an index already exists, increase it by one.
}
