% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-importData-prepareData.R
\name{joinColumnsUI}
\alias{joinColumnsUI}
\alias{joinColumnsServer}
\title{Join Columns UI}
\usage{
joinColumnsUI(id)

joinColumnsServer(id, preparedData)
}
\arguments{
\item{id}{id of module}

\item{preparedData}{(reactive) (reactive) selected data, possibly already modified}
}
\description{
UI of the module

Server function of the module
}
