% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-DataProcessLink-class.R
\name{load_data_from_link.DataProcessLink}
\alias{load_data_from_link.DataProcessLink}
\title{S3 method: Load data from DataProcessLink
Loads data from the source specified in the DataProcessLink object.}
\usage{
\method{load_data_from_link}{DataProcessLink}(object, ...)
}
\arguments{
\item{object}{The DataProcessLink object to load data from.}

\item{...}{Additional arguments, including:
- values: A reactiveValues object to store loading results.
- customNames: A list with custom naming options (withRownames, withColnames).}
}
\value{
A reactiveValues object containing the loaded data and status.
}
\description{
S3 method: Load data from DataProcessLink
Loads data from the source specified in the DataProcessLink object.
}
