% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-importData-selectSource.R
\name{loadCKANResourceUI}
\alias{loadCKANResourceUI}
\alias{loadCKANResourceServer}
\title{Load CKAN Resource UI}
\usage{
loadCKANResourceUI(id)

loadCKANResourceServer(id)
}
\arguments{
\item{id}{id of module}
}
\description{
UI of the module

Server function of the module
}
