% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-importData.R
\name{loadDataWrapper}
\alias{loadDataWrapper}
\title{Load Data Wrapper}
\usage{
loadDataWrapper(
  values,
  filepath,
  filename,
  colNames,
  type,
  sep,
  dec,
  withRownames,
  withColnames,
  includeSd,
  outputAsMatrix,
  sheetId
)
}
\arguments{
\item{values}{(list) list with import specifications}

\item{filepath}{(character) url or path}

\item{filename}{(character) url or file name}

\item{colNames}{(reactive) use this for colnames of imported data}

\item{type}{(character) file type input}

\item{sep}{(character) column separator input}

\item{dec}{(character) decimal separator input}

\item{withRownames}{(logical) contains rownames input}

\item{includeSd}{(logical) include sd input}

\item{outputAsMatrix}{(logical) TRUE if output must be a matrix,
e.g. for batch = TRUE in Resources}

\item{sheetId}{(numeric) sheet id}
}
\description{
Load Data Wrapper
}
