% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-loadImport.R
\name{loadImport}
\alias{loadImport}
\title{Load Import}
\usage{
loadImport(importType, params, expectedFileInZip)
}
\arguments{
\item{importType}{(character) type of import, either "data" or "model" or "zip" or "list".
ImportType == "model" expects a zip file containing a model. The file will be unzipped,
the model object extracted, and checked if it is valid for the app.
ImportType == "zip" enables the optional parameter 'expectedFileInZip'. The file is validated
and the path to the zip file will be returned.}

\item{params}{(list) named list of parameters required to import data or a model}

\item{expectedFileInZip}{(character) (optional) This parameter is ignored if importType != "zip".
File names that must be contained in the zip upload.}
}
\description{
Wrapper to load "data", "model", "zip" or "list" file.
}
