% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-loadList.R
\name{loadList}
\alias{loadList}
\title{Load List}
\usage{
loadList(path, type = "json")
}
\arguments{
\item{path}{(character) path to the file}

\item{type}{(character) file type input}
}
\description{
Load a list from a file. Currently only supports `"json"` files.
}
