% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-loadList.R
\name{loadListWrapper}
\alias{loadListWrapper}
\title{Load List Wrapper}
\usage{
loadListWrapper(values, filepath, type = "json")
}
\arguments{
\item{values}{(list) list with import specifications}

\item{filepath}{(character) url or path}

\item{type}{(character) file type input}
}
\description{
Wrapper to load a list from a `"json"` file.
}
