% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-loadModel.R
\name{loadModel}
\alias{loadModel}
\title{Load Model}
\usage{
loadModel(
  filepath,
  subFolder,
  rPackageName,
  onlySettings,
  fileExtension = "zip"
)
}
\arguments{
\item{filepath}{(character) path to the model file}

\item{subFolder}{(character) (optional) subfolder containing loadable .zip files}

\item{rPackageName}{(character) name of the package (as in the Description file) in which this
module is called. This value is needed to determine the Github repository of the package for
loading example models or zip or data links.}

\item{onlySettings}{(logical) if TRUE allow only upload of user inputs and user data}

\item{fileExtension}{(character) (optional) app specific file extension, e.g. "resources",
"bpred", "bmsc"}
}
\description{
Load Model
}
