% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-loadZip.R
\name{loadZipWrapper}
\alias{loadZipWrapper}
\title{Load Zip Wrapper}
\usage{
loadZipWrapper(filepath, fileExtension = "zip")
}
\arguments{
\item{filepath}{(character) path to the model file}

\item{fileExtension}{(character) (optional) app specific file extension, e.g. "resources",
"bpred", "bmsc"}
}
\description{
Load Zip Wrapper
}
