% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-importData-mergeData.R
\name{matchColClasses}
\alias{matchColClasses}
\title{Match Column Classes}
\usage{
matchColClasses(
  df1,
  df2,
  xColNames,
  yColNames,
  df1Id = "table1",
  isTest = FALSE
)
}
\arguments{
\item{df1}{(data.frame) data frame 1}

\item{df2}{(data.frame) data frame 2}

\item{xColNames}{(character) column names of data frame 1}

\item{yColNames}{(character) column names of data frame 2}

\item{df1Id}{(character) id of data frame 1}

\item{isTest}{(logical) if TRUE, no warning is shown}
}
\description{
Match Column Classes
}
