% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-importData-mergeData.R
\name{mergeDataUI}
\alias{mergeDataUI}
\alias{mergeDataServer}
\title{Merge Data UI}
\usage{
mergeDataUI(id)

mergeDataServer(id, dataProcessList)
}
\arguments{
\item{id}{id of module}

\item{dataProcessList}{(list) list of data to be merged}
}
\description{
UI of the merge data module

Server function of the merge data module
}
