% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-importData.R
\name{modelImportOptions}
\alias{modelImportOptions}
\title{Model Import Options}
\usage{
modelImportOptions(
  fileExtension = "zip",
  mainFolder = "predefinedModels",
  subFolder = NULL,
  rPackageName = "",
  onlySettings = FALSE
)
}
\arguments{
\item{fileExtension}{(character) (otional) app specific file extension, e.g. "resources", "bmsc",
"bpred", or (app-unspecific) "zip". Only files with this extension are valid for import.}

\item{mainFolder}{(character) folder containing all loadable .zip files.
This parameter is ignored if importType == "data"}

\item{subFolder}{(character) (optional) subfolder containing loadable .zip files.
This parameter is ignored if importType == "data"}

\item{rPackageName}{(character) If not NULL, than the uploaded file must be a downloaded file
from this R package. This parameter is ignored if importType == "data"}

\item{onlySettings}{(logical) if TRUE allow only upload of user inputs and user data.
This parameter is ignored if importType == "data"}
}
\value{
list of extra options for \code{impoertDataServer()} that are relevant if
\code{importType = "model"}
}
\description{
Model Import Options
}
