% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-DataProcessLink-class.R
\name{new_DataProcessLink}
\alias{new_DataProcessLink}
\title{Constructor for DataProcessLink
Creates a new DataProcessLink object.}
\usage{
new_DataProcessLink(input, filename, unprocessed = TRUE, history = list())
}
\arguments{
\item{input}{A list containing file, source, and query inputs.}

\item{filename}{The name of the file associated with this data item.}

\item{unprocessed}{Logical indicating if the data is unprocessed, currently TRUE for data links.}

\item{history}{Optional list of history entries.}
}
\value{
A new DataProcessLink object.
}
\description{
Constructor for DataProcessLink
Creates a new DataProcessLink object.
}
