% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-DownloadBundle-class.R
\name{new_DownloadBundle}
\alias{new_DownloadBundle}
\title{Create a new DownloadBundle object}
\usage{
new_DownloadBundle(
  package_name,
  sub_model = NULL,
  help_html = "",
  compression_level = 9
)
}
\arguments{
\item{package_name}{The name of the R package for which the bundle is created.}

\item{sub_model}{Optional sub-model name within the R package.}

\item{help_html}{Optional HTML content for help documentation.}

\item{compression_level}{The level of compression to use when zipping the bundle (default is 9).}
}
\value{
A DownloadBundle object containing the specified parameters and a temporary directory for file storage.
}
\description{
This function creates a new DownloadBundle object, which is used to
prepare a bundle of files for download, including RDS files, notes, and help files.
}
