% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-importData-linkToData.R
\name{observeUploadDataLink}
\alias{observeUploadDataLink}
\title{Observe Upload of a link to import data}
\usage{
observeUploadDataLink(
  id,
  input,
  output,
  session,
  isInternet,
  dataSource,
  customNames,
  mergeList
)
}
\arguments{
\item{id}{module id}

\item{input}{input object from server function}

\item{output}{output object from server function}

\item{session}{session from server function}

\item{isInternet}{(logical) set TRUE, if there is an internet connection. This parameter is
ignored if \code{type = "file"} or \code{type = "remoteModel"}}

\item{dataSource}{(reactiveValues) path, filename, type and input, output of \code{selectSourceServer()}}

\item{customNames}{settings for custom column and row names}

\item{mergeList}{(reactiveVal) list of data imports submitted for data processing via buttons
'Create Query with data' or 'Prepare / Merge data'}
}
\description{
Observe Upload of a link to import data
}
