% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-config.R
\docType{data}
\name{pkg.env}
\alias{pkg.env}
\title{globals}
\format{
An object of class \code{environment} of length 1.
}
\usage{
pkg.env
}
\description{
Contains the package's global variables
}
\keyword{datasets}
