% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-downAndUpload.R
\name{prefixSysTime}
\alias{prefixSysTime}
\title{Prefix Systime}
\usage{
prefixSysTime(fileString)
}
\arguments{
\item{fileString}{(character) file name}
}
\value{
(character) file name with prefix of current time
}
\description{
Prefixes a file name with the current time
}
