% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-importData-prepareData.R
\name{prepareDataUI}
\alias{prepareDataUI}
\alias{prepareDataServer}
\title{Prepare Data UI}
\usage{
prepareDataUI(id)

prepareDataServer(id, dataProcessList)
}
\arguments{
\item{id}{id of module}

\item{dataProcessList}{(list) list of selected data}
}
\description{
UI of the module

Server function of the module
}
