% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-importData-previewData.R
\name{previewDataUI}
\alias{previewDataUI}
\alias{previewDataServer}
\title{Preview Data UI}
\usage{
previewDataUI(id, title = "Preview data")

previewDataServer(id, dat)
}
\arguments{
\item{id}{id of module}

\item{title}{title}

\item{dat}{(reactive) data.frame of preview data to be displayed}
}
\description{
UI of the module

Server function of the module
}
