% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-importData-selectData.R
\name{previewDataUI}
\alias{previewDataUI}
\title{Preview Data UI}
\usage{
previewDataUI(id, title = "Preview data")
}
\arguments{
\item{id}{id of module}

\item{title}{title}
}
\description{
UI of the module
}
