% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-DataProcessItem-class.R
\name{print.DataProcessItem}
\alias{print.DataProcessItem}
\title{S3 method: Print DataProcessItem
Prints a summary of the DataProcessItem object.}
\usage{
\method{print}{DataProcessItem}(x, ...)
}
\arguments{
\item{x}{The DataProcessItem object to print.}

\item{...}{Additional arguments (not used).}
}
\description{
S3 method: Print DataProcessItem
Prints a summary of the DataProcessItem object.
}
