% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-DataProcessLink-class.R
\name{print.DataProcessLink}
\alias{print.DataProcessLink}
\title{S3 method: Print DataProcessLink
Prints a summary of the DataProcessLink object.}
\usage{
\method{print}{DataProcessLink}(x, ...)
}
\arguments{
\item{x}{The DataProcessLink object to print.}

\item{...}{Additional arguments (not used).}
}
\description{
S3 method: Print DataProcessLink
Prints a summary of the DataProcessLink object.
}
