% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-importData-queryData.R
\name{queryDataUI}
\alias{queryDataUI}
\alias{queryDataServer}
\title{Query Data UI}
\usage{
queryDataUI(id)

queryDataServer(id, dataProcessList, isActiveTab)
}
\arguments{
\item{id}{id of module}

\item{dataProcessList}{(list) list of data to be merged}

\item{isActiveTab}{(reactive) TRUE if tab of query module is the active}
}
\description{
UI of the query data module

Server function of the qery data module
}
