% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-remoteModels.R
\name{remoteModelsUI}
\alias{remoteModelsUI}
\title{Remote models module}
\usage{
remoteModelsUI(id)
}
\arguments{
\item{id}{id of module}
}
\description{
Displays a select input to choose a remote model and a button to load it
}
