% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-importData-prepareData.R
\name{renameColumnsServer}
\alias{renameColumnsServer}
\title{Rename Columns Server}
\usage{
renameColumnsServer(id, preparedData)
}
\arguments{
\item{id}{id of module}

\item{preparedData}{(reactive) selected data, possibly already modified}
}
\description{
Server function of the module
}
