% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-renameExistingNames.R
\name{renameExistingNames}
\alias{renameExistingNames}
\title{Rename Existing Names}
\usage{
renameExistingNames(newList, oldList)
}
\arguments{
\item{newList}{(list) list with named object(s) to be appended to an existing list}

\item{oldList}{(list) list to be extended}
}
\description{
Rename Existing Names
}
