% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-rgpt3.R
\name{rgpt_models_for_sql}
\alias{rgpt_models_for_sql}
\title{Make a request to the GPT API to extract the list of models}
\usage{
rgpt_models_for_sql(
  excludePattern = "babbage|curie|dall-e|davinci|text-embedding|tts|whisper"
)
}
\arguments{
\item{excludePattern}{A character vector that contains a regular expression that will be used to
exclude models from the list. By default, the function excludes models that are not designed to
obtain SQL queries.}
}
\value{
A character vector with the model IDs
}
\description{
`rgpt_models()` sends a request to the Open AI GPT API to extract the list of
models that are available for chat completions. This function is useful to
check which models are available for your completions.
}
