% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-rgpt3.R
\name{rgpt_test_completion}
\alias{rgpt_test_completion}
\title{Make a test request to the GPT API}
\usage{
rgpt_test_completion(verbose = T)
}
\arguments{
\item{verbose}{(boolean) if TRUE prints the actual prompt and GPT completion 
of the test request (default: TRUE).}
}
\value{
A message of success or failure of the connection.
}
\description{
`rgpt_test_completion()` sends a basic 
[completion request](https://beta.openai.com/docs/api-reference/completions) 
to the Open AI GPT API.
This function was copied from https://github.com/ben-aaron188/rgpt3.
}
\examples{
rgpt_test_completion()
}
