% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-importData-selectData.R
\name{selectDataServer}
\alias{selectDataServer}
\title{Select Data Server}
\usage{
selectDataServer(
  id,
  importType,
  mergeList,
  customNames,
  openPopupReset,
  internetCon,
  ckanFileTypes = c("xls", "xlsx", "csv", "odt", "txt"),
  ignoreWarnings = FALSE,
  rPackageName = "",
  onlySettings = FALSE,
  expectedFileInZip = c(),
  exampleOptions = exampleOptions()
)
}
\arguments{
\item{id}{id of module}

\item{importType}{(character) type of import, either "data" or "model" or "zip".
ImportType == "zip" enables the optional parameter 'expectedFileInZip'.}

\item{mergeList}{(list) list of selected data}

\item{customNames}{settings for custom column and row names}

\item{openPopupReset}{(reactive) if TRUE reset ckan source inputs}

\item{internetCon}{(reactive) TRUE if there is an internet connection}

\item{ckanFileTypes}{(character) file types allowed for import from Pandora ("ckan")}

\item{ignoreWarnings}{TRUE to enable imports in case of warnings}

\item{rPackageName}{(character) If not NULL, than the uploaded file must be a downloaded file
from this R package. This parameter is ignored if importType == "data"}

\item{onlySettings}{(logical) if TRUE allow only upload of user inputs and user data.
This parameter is ignored if importType == "data"}

\item{expectedFileInZip}{(character) (optional) This parameter is ignored if importType != "zip".
File names that must be contained in the zip upload.}

\item{exampleOptions}{extra options for remote or local examples}
}
\description{
Server function of the module
}
