% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-importData-selectFileType.R
\name{selectFileTypeServer}
\alias{selectFileTypeServer}
\title{Select File Type Server}
\usage{
selectFileTypeServer(id, dataSource)
}
\arguments{
\item{id}{id of module}

\item{dataSource}{(reactive) data source from selectSourceServer()}
}
\description{
Server function of the module
}
