% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-importData-selectFileType.R
\name{selectFileTypeUI}
\alias{selectFileTypeUI}
\title{Select File Type UI}
\usage{
selectFileTypeUI(
  id,
  defaultFileTypes = config()[["dataFileTypes"]],
  userFileTypes = c()
)
}
\arguments{
\item{id}{id of module}

\item{defaultFileTypes}{(character) default file types}

\item{userFileTypes}{(character) user file types specified in "Pandora Platform" settings}
}
\description{
UI of the module
}
