% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-importData-selectData.R
\name{selectImportParams}
\alias{selectImportParams}
\title{Select Import params}
\usage{
selectImportParams(importType, params)
}
\arguments{
\item{importType}{(character) type of import, either "data" or "model"}

\item{params}{(list) named list of parameters required to import data or a model}
}
\value{
(list) named list of parameters required for the particular importType
}
\description{
Select Import params
}
