% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-importData-selectData.R
\name{selectSourceServer}
\alias{selectSourceServer}
\title{Select Source Server}
\usage{
selectSourceServer(id)
}
\arguments{
\item{id}{id of module}
}
\description{
Server function of the module
}
