% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-importModule.R
\name{setImportTitle}
\alias{setImportTitle}
\title{Import title}
\usage{
setImportTitle(title = "", importType, version)
}
\arguments{
\item{title}{(character) title of the import dialog}

\item{importType}{(character) type of import, either "data", "model", "zip" or "list".}

\item{version}{(character) version of the package}
}
\description{
Import title
}
