% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-setModuleTitle.R
\name{setModuleTitle}
\alias{setModuleTitle}
\title{Set Module Title}
\usage{
setModuleTitle(title, titleTag = "h4")
}
\arguments{
\item{title}{(character) module title}

\item{titleTag}{(character) HTML tag to put around the title, e.g. "h4" for \code{h4} from
\code{htmltools}}
}
\description{
Set Module Title
}
