% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-showInfoUI.R
\name{showInfoUI}
\alias{showInfoUI}
\title{Show Info UI}
\usage{
showInfoUI(infoText, color = "#0072B2")
}
\arguments{
\item{infoText}{charackter info text to show when hoovering}

\item{color}{(character) text color}
}
\description{
Show Info UI
}
