% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-importData-prepareData.R
\name{splitColumnsUI}
\alias{splitColumnsUI}
\alias{splitColumnsServer}
\title{Split Columns UI}
\usage{
splitColumnsUI(id)

splitColumnsServer(id, preparedData)
}
\arguments{
\item{id}{id of module}

\item{preparedData}{(reactive) (reactive) selected data, possibly already modified}
}
\description{
UI of the module

Server function of the module
}
