% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-importData-storeInputs.R
\name{storeInputsServer}
\alias{storeInputsServer}
\title{Store Inputs Server}
\usage{
storeInputsServer(id, inputs, githubRepo, rPackageName)
}
\arguments{
\item{id}{id of module}

\item{inputs}{(reactive) user inputs}

\item{githubRepo}{(character) name of used github repository, e.g. "bpred"}

\item{rPackageName}{(character) name of the package (as in the description file) in which this
module is applied, e.g. "mpiBpred"}
}
\description{
Server function of the module
}
