% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-toolsPanel.R
\name{toolsImportServer}
\alias{toolsImportServer}
\title{Server function of toolsImport module}
\usage{
toolsImportServer(id, defaultSource = "ckan")
}
\arguments{
\item{id}{module id}

\item{defaultSource}{(character) default source for input "Source", e.g. "ckan", "file", or "url"}
}
\description{
Server function of toolsImport module
}
