% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-toolsPanel.R
\name{toolsLoadServer}
\alias{toolsLoadServer}
\title{Server function of toolsLoad module}
\usage{
toolsLoadServer(id)
}
\arguments{
\item{id}{module id}
}
\description{
Server function of toolsLoad module
}
