% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-toolsPanelModule.R
\name{toolsPanelServer}
\alias{toolsPanelServer}
\title{Server function of toolsPanel module}
\usage{
toolsPanelServer(id)
}
\arguments{
\item{id}{module id}
}
\description{
Server function of toolsPanel module
}
