% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-tryCatchWithErrorsAndWarnings.R
\name{tryCatchWithWarningsAndErrors}
\alias{tryCatchWithWarningsAndErrors}
\title{Try Catch with Warnings and Errors}
\usage{
tryCatchWithWarningsAndErrors(
  expr,
  errorTitle = "Modeling failed",
  warningTitle = "",
  alertStyle = "shinyjs",
  silent = FALSE
)
}
\arguments{
\item{expr}{expression to be evaluated.}

\item{errorTitle}{(character) error message title.}

\item{warningTitle}{(character) warning message title.}

\item{alertStyle}{(character) Either "shinyjs", or "shinyalert". Specifies how an error or a
warning is given out. If "shinyjs" than shinyjs::alert is used; if "shinyalert" than
shinyalert::shinyalert is used.}

\item{silent}{(logical) if TRUE prevents an alert, instead a warning will be displayed. Use this
when not applied within a Shiny session.}
}
\description{
Catches errors and warnings and forwards them as an alert. If an error occurs, NULL is returned.
If a warning occurs, the result is returned.
}
