% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-DataProcessItem-class.R
\name{update.DataProcessItem}
\alias{update.DataProcessItem}
\title{S3 method: Update fields of a DataProcessItem
Updates fields of a DataProcessItem object.}
\usage{
\method{update}{DataProcessItem}(object, ...)
}
\arguments{
\item{object}{The DataProcessItem object to update.}

\item{...}{New args to update. Possible args: data, input, unprocessed, filename, history.}
}
\value{
The updated DataProcessItem object.
}
\description{
S3 method: Update fields of a DataProcessItem
Updates fields of a DataProcessItem object.
}
