% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-importData-loadImport.R
\name{updateMessage}
\alias{updateMessage}
\title{Update Message}
\usage{
updateMessage(dat, element = c("data", "inputs", "model"), msgString)
}
\arguments{
\item{dat}{(list) list containing data, results and warning messages of import}

\item{element}{(character) list object, one of "data", "inputs", "model"}

\item{msgString}{(character) description of element used in messages}
}
\description{
Update Message
}
