% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-updateListNamesIfDuplicate.R
\name{updateNameEntryIfDuplicate}
\alias{updateNameEntryIfDuplicate}
\title{Update Names Entries if Duplicate}
\usage{
updateNameEntryIfDuplicate(newList, oldList, listType = "List", isShiny = TRUE)
}
\arguments{
\item{newList}{(list) The names of this list should be updated}

\item{oldList}{(list) The names of this list should be kept}

\item{listType}{(character) type of list, will be used in the notification if names had been
changed.}

\item{isShiny}{(logical) FALSE if not used in a shiny app, e.g. in tests.}
}
\description{
Updates the entries called "name" of a list (newList), if the names already exist in another list
(oldList).
This is useful if two lists should be concatenated and duplicated names must be prevented.
}
