% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-importData-linkToData.R
\name{updateUserInputs}
\alias{updateUserInputs}
\title{Update User Inputs}
\usage{
updateUserInputs(id, input, output, session, userInputs, inDataTools = FALSE)
}
\arguments{
\item{id}{module id}

\item{input}{input object from server function}

\item{output}{output object from server function}

\item{session}{session from server function}

\item{userInputs}{(list) list of inputs to be updated}

\item{inDataTools}{(logical) internal}
}
\description{
Update values from 'input' with all values from userInputs. Entries that are NULL are removed.
Only updates entries that are present in 'input'. Copy of \code{shinyTools::updateUserInputs()}
to avoid dependency to another helper package.
Export from 'DataTools' will be removed in future versions.
}
