% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-modelUpload.R
\name{uploadModelServer}
\alias{uploadModelServer}
\title{Server function for Upload Model}
\usage{
uploadModelServer(id, githubRepo)
}
\arguments{
\item{id}{id of module}

\item{githubRepo}{(character) name of the app's github repository, currently "bpred",
"osteo-bior", "resources"}
}
\description{
Backend for upload model module
}
