% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-downAndUpload.R
\name{uploadModelUI}
\alias{uploadModelUI}
\title{Upload model module}
\usage{
uploadModelUI(id, label, width = NULL)
}
\arguments{
\item{id}{id of module}

\item{label}{label of module}
}
\description{
UI function to upload a zip file with exportNotes and a list of models
}
