% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-importData.R
\name{zipImportOptions}
\alias{zipImportOptions}
\title{Zip Import Options}
\usage{
zipImportOptions(expectedFileInZip = c())
}
\arguments{
\item{expectedFileInZip}{(character) (optional) This parameter is ignored if importType != "zip".
File names that must be contained in the zip upload.}
}
\value{
list of extra options for \code{impoertDataServer()} that are relevant if
\code{importType = "data"}
}
\description{
Zip Import Options
}
