% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-downAndUpload.R
\name{addHelpFile}
\alias{addHelpFile}
\title{Add Help File}
\usage{
addHelpFile(filesToZip, helpHTML, tempDir)
}
\arguments{
\item{filesToZip}{(character) files to be zipped}

\item{helpHTML}{content of help function}

\item{tempDir}{(character) temporary directory to store the help file}
}
\value{
(character) updated paths to files to zip
}
\description{
Add Help File
}
