% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-downAndUpload.R
\name{addNotesFile}
\alias{addNotesFile}
\title{Add Notes File}
\usage{
addNotesFile(filesToZip, notes, tempDir)
}
\arguments{
\item{filesToZip}{(character) files to be zipped}

\item{notes}{(character) notes to be added to the zip file}

\item{tempDir}{(character) temporary directory to store the notes file}
}
\value{
(character) updated paths to files to zip
}
\description{
Add Notes File
}
