% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-importData-configureData.R
\name{configureDataUI}
\alias{configureDataUI}
\alias{configureDataServer}
\title{Configure Data UI}
\usage{
configureDataUI(
  id,
  batch,
  outputAsMatrix,
  isLink = FALSE,
  customHelpText = importOptions()[["customHelpText"]],
  defaultFileTypes = config()[["dataFileTypes"]],
  userFileTypes = c()
)

configureDataServer(
  id,
  dataProcessList,
  customNames,
  dataSource,
  dataSourceInputs,
  ignoreWarnings = FALSE
)
}
\arguments{
\item{id}{id of module}

\item{batch}{(logical) use batch import.}

\item{outputAsMatrix}{(logical) TRUE if output must be a matrix,
e.g. for batch = TRUE in Resources.}

\item{isLink}{(logical) if TRUE, the data source is a link}

\item{customHelpText}{(list) A help text element that can be added to a UI definition. Output of
\code{shiny::helpText(...)}.}

\item{defaultFileTypes}{(character) default file types}

\item{userFileTypes}{(character) user file types specified in "Pandora Platform" settings}

\item{dataProcessList}{(reactiveVal) list of data imports submitted for data processing via buttons
'Create Query with data' or 'Prepare / Merge data'}

\item{customNames}{settings for custom column and row names}

\item{dataSource}{(reactiveValues) path, filename, type and input, output of \code{selectSourceServer()}}

\item{dataSourceInputs}{(reactiveValues) inputs related to the data source}

\item{ignoreWarnings}{(logical) TRUE to enable imports in case of warnings}
}
\description{
UI of the module

Server function of the module
}
