% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-downAndUpload.R
\name{downloadModelServer}
\alias{downloadModelServer}
\title{Server function download model}
\usage{
downloadModelServer(
  id,
  dat,
  inputs,
  model,
  rPackageName,
  modelSubFolder = NULL,
  helpHTML = "",
  onlySettings = FALSE,
  compress = TRUE,
  compressionLevel = 9
)
}
\arguments{
\item{id}{namespace id}

\item{dat}{(reactive) user data}

\item{inputs}{(reactive) user inputs}

\item{model}{(reactive) model output object}

\item{rPackageName}{(character) name of the package (as in the description file) in which this
module is applied, e.g. "mpiBpred"}

\item{modelSubFolder}{(character) possible subfolder containing predefined models}

\item{helpHTML}{content of help function}

\item{onlySettings}{(logical) if TRUE allow only download of user inputs and user data}

\item{compress}{a logical specifying whether saving to a named file is to use "gzip" compression,
or one of "gzip", "bzip2" or "xz" to indicate the type of compression to be used. Ignored if
file is a connection.}

\item{compressionLevel}{A number between 1 and 9. 9 compresses best, but it also takes the
longest.}
}
\description{
Backend for download model module
}
