% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-downAndUpload.R
\name{downloadModelServer}
\alias{downloadModelServer}
\title{Server function download model}
\usage{
downloadModelServer(
  id,
  dat,
  inputs,
  model,
  rPackageName,
  onlySettings = reactive(FALSE),
  compress = TRUE
)
}
\arguments{
\item{id}{namespace id}

\item{dat}{(reactive) user data}

\item{inputs}{(reactive) user inputs}

\item{model}{(reactive) model output object}

\item{rPackageName}{(character) name of the package (as in the description file) in which this
module is applied, e.g. "mpiBpred"}

\item{onlySettings}{(logical) if TRUE allow only download of user inputs and user data}

\item{compress}{a logical specifying whether saving to a named file is to use "gzip" compression,
or one of "gzip", "bzip2" or "xz" to indicate the type of compression to be used. Ignored if
file is a connection.}
}
\description{
Backend for download model module
}
