% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-modelDownload.R
\name{downloadModelServer}
\alias{downloadModelServer}
\title{Server function download model}
\usage{
downloadModelServer(
  id,
  listOfModels,
  uploadedNotes,
  modelOutputTag = "fit",
  help = NULL
)
}
\arguments{
\item{id}{namespace id}

\item{listOfModels}{(reactive) list of model objects}

\item{uploadedNotes}{(reactive) variable that stores content of README.txt}

\item{modelOutputTag}{(character) name of the model output in the model object, e.g. "fit"}

\item{help}{(html) html containing information, e.g. getHelp(input$tab)}
}
\description{
Backend for download model module
}
