% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-modelDownload.R
\name{downloadModelUI}
\alias{downloadModelUI}
\title{Download model module}
\usage{
downloadModelUI(
  id,
  label = "Download Model",
  emptyChoicesText = "Create models ..."
)
}
\arguments{
\item{id}{id of module}

\item{label}{label of module}

\item{emptyChoicesText}{placeholder text if no models are available}
}
\description{
UI function to download a zip file with notes and a list of models
}
