% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-downAndUpload.R
\name{downUploadButtonServer}
\alias{downUploadButtonServer}
\title{downUploadButton module server}
\usage{
downUploadButtonServer(
  id,
  dat,
  inputs,
  model,
  rPackageName,
  githubRepo,
  mainFolder = "predefinedModels",
  subFolder = NULL,
  helpHTML = "",
  modelNotes = reactive(""),
  onlySettings = FALSE,
  compress = TRUE,
  compressionLevel = 9,
  reset = reactive(FALSE),
  title = "Download and Upload of Models",
  labelRemote = "Load online model",
  labelLocal = "Load local model"
)
}
\arguments{
\item{id}{module id}

\item{dat}{(reactive) user data}

\item{inputs}{(reactive) user inputs}

\item{model}{(reactive) model output object}

\item{rPackageName}{(character) name of the package (as in the description file) in which this
module is applied, e.g. "mpiBpred"}

\item{githubRepo}{(character) name of used github repository, e.g. "bpred"}

\item{mainFolder}{(character) folder containing all loadable .zip files}

\item{subFolder}{(character) name of the (optional) subfolder containing loadable .zip files}

\item{helpHTML}{content of help function}

\item{modelNotes}{(reactive) notes regarding the object to be saved and displayed when uploaded}

\item{onlySettings}{(logical) if TRUE allow only download of user inputs and user data}

\item{compress}{a logical specifying whether saving to a named file is to use "gzip" compression,
or one of "gzip", "bzip2" or "xz" to indicate the type of compression to be used. Ignored if
file is a connection.}

\item{compressionLevel}{A number between 1 and 9. 9 compresses best, but it also takes the
longest.}

\item{reset}{(reactive) resets the selection of the online files}

\item{title}{title used inside the modal}
}
\description{
downUploadButton module server
}
