% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-downAndUpload.R
\name{downUploadButtonServer}
\alias{downUploadButtonServer}
\title{downUploadButton module server}
\usage{
downUploadButtonServer(
  id,
  dat,
  inputs,
  model,
  rPackageName,
  helpHTML = "",
  onlySettings = FALSE,
  compress = TRUE,
  compressionLevel = 9,
  githubRepo,
  folderOnGithub = "/predefinedModels",
  reset = reactive(FALSE)
)
}
\arguments{
\item{id}{module id}

\item{dat}{(reactive) user data}

\item{inputs}{(reactive) user inputs}

\item{model}{(reactive) model output object}

\item{rPackageName}{(character) name of the package (as in the description file) in which this
module is applied, e.g. "mpiBpred"}

\item{helpHTML}{content of help function}

\item{onlySettings}{(logical) if TRUE allow only download of user inputs and user data}

\item{compress}{a logical specifying whether saving to a named file is to use "gzip" compression,
or one of "gzip", "bzip2" or "xz" to indicate the type of compression to be used. Ignored if
file is a connection.}

\item{compressionLevel}{A number between 1 and 9. 9 compresses best, but it also takes the
longest.}

\item{githubRepo}{(character) name of used github repository, e.g. "bpred"}

\item{folderOnGithub}{(character) folder on github where remote models are stored. This should
correspond to 'pathToLocal' since online and offline models should be the same and up-to-date}

\item{reset}{(reactive) resets the selection of the online model}
}
\description{
downUploadButton module server
}
