% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-downAndUpload.R
\name{downUploadButtonServer}
\alias{downUploadButtonServer}
\title{downUploadButton module server}
\usage{
downUploadButtonServer(
  id,
  dat,
  inputs,
  model,
  rPackageName,
  githubRepo,
  mainFolder = "predefinedModels",
  subFolder = NULL,
  fileExtension = "zip",
  helpHTML = "",
  modelNotes = reactive(""),
  onlySettings = FALSE,
  compress = TRUE,
  compressionLevel = 9,
  reset = reactive(FALSE),
  title = "Download and Upload of Models",
  labelRemote = "Load online model",
  labelLocal = "Load local model"
)
}
\arguments{
\item{id}{module id}

\item{dat}{(reactive) user data}

\item{inputs}{(reactiveValues) reactiveValues list of user inputs, in most cases just the "inputs" list}

\item{model}{(reactive) model output object}

\item{rPackageName}{(character) name of the package (as in the Description file) in which this
module is called. If not NULL, than the uploaded file must be a downloaded file
from the R package where \code{importDataServer} was called. This parameter is ignored if
\code{importType == "data"}.}

\item{githubRepo}{(character) name of used github repository, e.g. "bpred"}

\item{mainFolder}{(character) folder containing all loadable .zip files. For most apps this
is folder "predefinedModels". In most apps it can be found under "inst/app/".}

\item{subFolder}{(character) (optional) subfolder containing loadable .zip files}

\item{fileExtension}{(character) (optional) app specific file extension, e.g. "resources",
"bpred", "bmsc"}

\item{helpHTML}{content of help function}

\item{modelNotes}{(reactive) notes regarding the object to be saved and displayed when uploaded}

\item{onlySettings}{(logical) if TRUE allow only download of user inputs and user data}

\item{compress}{a logical specifying whether saving to a named file is to use "gzip" compression,
or one of "gzip", "bzip2" or "xz" to indicate the type of compression to be used. Ignored if
file is a connection.}

\item{compressionLevel}{A number between 1 and 9. 9 compresses best, but it also takes the
longest.}

\item{reset}{(reactive) resets the selection of the online files}

\item{title}{title used inside the modal}

\item{labelRemote}{label of the input for remote files}

\item{labelLocal}{label of the input for local files}
}
\description{
downUploadButton module server
}
