% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-importData-loadImport.R
\name{extractValuesFromModel}
\alias{extractValuesFromModel}
\title{Extract Values From Model}
\usage{
extractValuesFromModel(importedModel, values, filename)
}
\arguments{
\item{importedModel}{(list) output of loadModel()}

\item{values}{(reactiveValues) empty list of values in the format of the output of loadDataWrapper}
}
\value{
(list) list of values in the format of the output of loadDataWrapper
}
\description{
Extract Values From Model
}
