% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-importData-selectSource.R
\name{filterCKANResourceServer}
\alias{filterCKANResourceServer}
\title{Filter CKAN Resource Server}
\usage{
filterCKANResourceServer(id)
}
\arguments{
\item{id}{id of module}
}
\description{
Server function of the module
}
