% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-importData-selectSource.R
\name{filterCKANResourceUI}
\alias{filterCKANResourceUI}
\title{Filter CKAN Resource UI}
\usage{
filterCKANResourceUI(id, ckanFileTypes)
}
\arguments{
\item{id}{id of module}

\item{ckanFileTypes}{(character) file types allowed for import from Pandora ("ckan"). E.g. for
`importType = "data"`: c("xls", "xlsx", "csv", "odt", "txt"); for `importType = "zip"`: c("zip");
 for `importType = "list"`: c("json")}
}
\description{
UI of the module
}
