% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-ckan.R
\name{getCKANGroupChoices}
\alias{getCKANGroupChoices}
\title{Get CKAN Group Choices}
\usage{
getCKANGroupChoices(groupList = data.frame())
}
\description{
Get choices that will be available in the ckanGroup (network) input
}
