% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-ckan.R
\name{getCKANRecordChoices}
\alias{getCKANRecordChoices}
\title{Get CKAN Record Choices}
\usage{
getCKANRecordChoices(network = "", pattern = "", packageList = data.frame())
}
\arguments{
\item{network}{(character) name of a Pandora network, e.g. an entry of the output from
\code{getNetworks()$name}}

\item{pattern}{(character) string for meta information search}

\item{packageList}{(data.frame) optional, output of callAPI() e.g. from a previous call to the
Pandora API.}
}
\description{
Get choices that will be available in the ckanRecord (repository) input
}
