% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-ckan.R
\name{getCKANResourcesChoices}
\alias{getCKANResourcesChoices}
\title{Get CKAN Resource Choices}
\usage{
getCKANResourcesChoices(
  fileType = character(),
  repository = "",
  network = "",
  pattern = ""
)
}
\arguments{
\item{fileType}{(character) list of relevant file types, e.g. c("xls", "xlsx", "csv", "odt")}

\item{repository}{(character) name of Pandora repository}

\item{network}{(character) name of Pandora network}

\item{pattern}{(character) string for meta information search}
}
\description{
Get choices that will be available in the ckanResource input
}
