% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-ckan.R
\name{getCKANResourcesChoices}
\alias{getCKANResourcesChoices}
\title{Get CKAN Resource Choices}
\usage{
getCKANResourcesChoices(ckanResources, types, sort = TRUE)
}
\arguments{
\item{ckanResources}{(list) output of `getCKANFiles() %>% filterCKANFileList()` for a specific record}

\item{types}{(character) user selected types to show}

\item{sort}{(logical) if TRUE sort choices alphabetically}
}
\description{
Select, filter and sort choices to be available in the ckanResource input
}
