% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-importData-selectData.R
\name{getDataSource}
\alias{getDataSource}
\title{Get Data Source}
\usage{
getDataSource(
  dataSource = list(file = NULL, filename = NULL, type = NULL),
  importType,
  input,
  type = c("ckan", "file", "url", "remoteModel"),
  isInternet = TRUE
)
}
\arguments{
\item{dataSource}{(reactiveValues)}

\item{importType}{(character) type of import, either "data" or "model" or "zip".
ImportType == "zip" enables the optional parameter 'expectedFileInZip'.}

\item{input}{(reactiveValues)}

\item{type}{(character) source of import, one of "ckan", "file", "url", "remoteModel".
Possible sources for data are: "ckan", "file", "url".
Possible sources for models are: "ckan", "file", "url", "remoteModel".}

\item{isInternet}{(logical) set TRUE, if there is an internet connection. This parameter is
ignored if \code{type = "file"} or \code{type = "remoteModel"}}
}
\description{
Get Data Source
}
