% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-importData-selectData.R
\name{getDataSource}
\alias{getDataSource}
\title{Get Data Source}
\usage{
getDataSource(
  importType,
  dataSource,
  input,
  type = c("ckan", "file", "url", "model"),
  isInternet = TRUE
)
}
\arguments{
\item{importType}{(character) type of import, either "data" or "model" or "zip".
ImportType == "zip" enables the optional parameter 'expectedFileInZip'.}

\item{dataSource}{(reactiveValues)}

\item{input}{(reactiveValues)}

\item{type}{(character) source of import, one of "ckan", "file", "url", "model".
Possible sources for data are: "ckan", "file", "url".
Possible sources for models are: "ckan", "file", "url", "model".}

\item{isInternet}{(logical) set TRUE, if there is an internet connection. This parameter is
ignored if \code{type = "file"} or \code{type = "model"}}
}
\description{
Get Data Source
}
