% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-modelUpload-remote.R
\name{getGithubContent}
\alias{getGithubContent}
\title{Get Github Content}
\usage{
getGithubContent(githubRepo)
}
\arguments{
\item{githubRepo}{(character) name of the app's github repository, currently "bpred",
"osteo-bior", "resources"}
}
\description{
Get content of api call to github folder
}
