% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-remoteModels.R
\name{getGithubContent}
\alias{getGithubContent}
\title{Get Github Content}
\usage{
getGithubContent(githubRepo, folderOnGithub = "/predefinedModels")
}
\arguments{
\item{githubRepo}{(character) name of used github repository, e.g. "bpred"}

\item{folderOnGithub}{(character) folder on github where remote models are stored. This should
correspond to 'pathToLocal' since online and offline models should be the same and up-to-date.}
}
\description{
Get content of api call to github folder
}
