% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-importData-selectData.R
\name{getGithubMapping}
\alias{getGithubMapping}
\title{Get Github Mapping}
\usage{
getGithubMapping(rPackage = "")
}
\arguments{
\item{rPackage}{(character) name of the R package (as in the Description file), must be empty or
specified in the config file of the package DataTools}
}
\description{
Maps the R package name to the respective Github repository
}
