% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-remoteModels.R
\name{getLocalModels}
\alias{getLocalModels}
\title{Get Local Models}
\usage{
getLocalModels(pathToLocal)
}
\arguments{
\item{pathToLocal}{(character) relative path to the folder storing local files}
}
\description{
Get Local Models
}
