% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-downAndUpload.R
\name{getPathToLocal}
\alias{getPathToLocal}
\title{Get Path to Local}
\usage{
getPathToLocal(mainFolder, subFolder, rPackageName = NULL)
}
\arguments{
\item{mainFolder}{(character) folder containing all loadable .zip files. For most apps this
is folder "predefinedModels". In most apps it can be found under "inst/app/".}

\item{subFolder}{(character) (optional) subfolder containing loadable .zip files}

\item{rPackageName}{(character) name of the package (as in the Description file) in which this
module is called. If not NULL, than the uploaded file must be a downloaded file
from the R package where \code{importDataServer} was called. This parameter is ignored if
\code{importType == "data"}.}
}
\description{
Get Path to Local
}
