% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-remoteModels.R
\name{getRemoteModelsFromGithub}
\alias{getRemoteModelsFromGithub}
\title{Get Remote Models From Github}
\usage{
getRemoteModelsFromGithub(githubRepo, rPackageName, rPackageVersion)
}
\arguments{
\item{githubRepo}{(character) name of used github repository, e.g. "bpred"}

\item{rPackageName}{(character) name of the package (as in the description file) where this
module is applied, e.g. "mpiBpred"}

\item{rPackageVersion}{(character) current version of the package where this module is applied,
e.g. packageVersion("mpiBpred")}
}
\description{
Get remote models from github directory
}
