% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-remoteModels.R
\name{getRemoteModelsFromGithub}
\alias{getRemoteModelsFromGithub}
\title{Get Remote Models From Github}
\usage{
getRemoteModelsFromGithub(githubRepo, folderOnGithub = "/predefinedModels")
}
\arguments{
\item{githubRepo}{(character) name of used github repository, e.g. "bpred"}

\item{folderOnGithub}{(character) folder on github where remote models are stored. This should
correspond to 'pathToLocal' since online and offline models should be the same and up-to-date}
}
\description{
Get remote models from github directory
}
