% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-remoteModels.R
\name{getRemoteModelsFromGithub}
\alias{getRemoteModelsFromGithub}
\title{Get Remote Models From Github}
\usage{
getRemoteModelsFromGithub(
  githubRepo,
  folderOnGithub = "/predefinedModels",
  fileExtension = "zip",
  apiOut = NULL
)
}
\arguments{
\item{githubRepo}{(character) name of used github repository, e.g. "bpred"}

\item{folderOnGithub}{(character) folder on github where remote files are stored. This should
correspond to 'pathToLocal' since online and offline files should be the same and up-to-date}

\item{fileExtension}{(character) (otional) app specific file extension, e.g. "resources", "bmsc",
"bpred", or (app-unspecific) "zip". Only files with this extension are valid for import.}

\item{apiOut}{output of `getGithubContent()` if it was already loaded}
}
\description{
Get remote models from github directory
}
