% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-importData-queryData.R
\name{gptUI}
\alias{gptUI}
\alias{gptServer}
\title{GPT UI}
\usage{
gptUI(id)

gptServer(id, autoCompleteList, isActiveTab)
}
\arguments{
\item{id}{id of module}

\item{autoCompleteList}{(list) word to be used for auto completion}

\item{isActiveTab}{(reactive) TRUE if tab of query module is the active}
}
\description{
UI of the gpt module

Server function of the gpt module
}
