% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-importData-queryData.R
\name{gptServer}
\alias{gptServer}
\title{GPT Server}
\usage{
gptServer(id, autoCompleteList)
}
\arguments{
\item{id}{id of module}

\item{autoCompleteList}{(list) word to be used for auto completion}
}
\description{
Server function of the gpt module
}
