% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-importData.R
\name{importDataServer}
\alias{importDataServer}
\title{Server function for data import}
\usage{
importDataServer(
  id,
  title = "",
  defaultSource = c("ckan", "file", "url", "remoteModel"),
  ckanFileTypes = c("xls", "xlsx", "csv", "odt", "txt"),
  ignoreWarnings = FALSE,
  importType = c("data", "model", "zip"),
  rowNames = reactiveVal(NULL),
  colNames = reactiveVal(NULL),
  customWarningChecks = list(),
  customErrorChecks = list(),
  batch = FALSE,
  outputAsMatrix = FALSE,
  fileExtension = "zip",
  mainFolder = NULL,
  subFolder = NULL,
  rPackageName = "",
  onlySettings = FALSE,
  expectedFileInZip = c(),
  options = importOptions()
)
}
\arguments{
\item{id}{namespace id}

\item{title}{title of data import module}

\item{defaultSource}{(character) default source for input "Source", e.g. "ckan", "file", or "url"}

\item{ckanFileTypes}{(character) file types allowed for import from Pandora ("ckan")}

\item{ignoreWarnings}{TRUE to enable imports in case of warnings}

\item{importType}{(character) type of import, either "data" or "model" or "zip".
ImportType == "zip" enables the optional parameter 'expectedFileInZip'.}

\item{rowNames}{(reactive) use this for rownames of imported data. This parameter is ignored if importType == "model"}

\item{colNames}{(reactive) use this for colnames of imported data. This parameter is ignored if importType == "model"}

\item{customWarningChecks}{list of reactive(!) functions which will be executed after importing
of data.
 functions need to return TRUE if check is successful or a character with a warning otherwise.
 This parameter is ignored if importType == "model"}

\item{customErrorChecks}{list of reactive(!) functions which will be executed after importing
of data.
  functions need to return TRUE if check is successful or a character with a warning otherwise.
  This parameter is ignored if importType == "model"}

\item{batch}{(logical) use batch import. This parameter is ignored if importType == "model"}

\item{outputAsMatrix}{(logical) TRUE if output must be a matrix,
e.g. for batch = TRUE in Resources. This parameter is ignored if importType == "model"}

\item{fileExtension}{(character) (otional) app specific file extension, e.g. "resources", "bmsc",
"bpred", or (app-unspecific) "zip". Only files with this extension are valid for import.}

\item{mainFolder}{(character) DEPRECATED. folder containing all loadable .zip files.
This parameter is ignored if importType == "data"}

\item{subFolder}{(character) (optional) subfolder containing loadable .zip files.
This parameter is ignored if importType == "data"}

\item{rPackageName}{(character) DEPRECATED. Instead, please use
\code{options = importOptions(rPackageName = <your package>)}.}

\item{onlySettings}{(logical) if TRUE allow only upload of user inputs and user data.
This parameter is ignored if importType == "data"}

\item{expectedFileInZip}{(character) (optional) This parameter is ignored if importType != "zip".
File names that must be contained in the zip upload.}

\item{options}{(list) Extra options for the import module. See \code{\link{importOptions}}.}
}
\description{
Backend for data import module
}
