% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-importData.R
\name{importDataServer}
\alias{importDataServer}
\title{Server function for data import}
\usage{
importDataServer(
  id,
  title = "",
  defaultSource = "ckan",
  ignoreWarnings = FALSE,
  importType = "data",
  rowNames = reactiveVal(NULL),
  colNames = reactiveVal(NULL),
  customWarningChecks = list(),
  customErrorChecks = list(),
  batch = FALSE,
  outputAsMatrix = FALSE,
  fileExtension = "zip",
  mainFolder = "predefinedModels",
  subFolder = NULL,
  rPackageName = "",
  onlySettings = FALSE,
  extractZipFun = NULL
)
}
\arguments{
\item{id}{namespace id}

\item{title}{title of data import module}

\item{defaultSource}{(character) default source for input "Source", e.g. "ckan", "file", or "url"}

\item{ignoreWarnings}{TRUE to enable imports in case of warnings}

\item{importType}{(character) type of import, either "data" or "model" or "zip".
ImportType == "zip" enables the optional parameter 'extractZipFun'.}

\item{rowNames}{(reactive) use this for rownames of imported data. This parameter is ignored if importType == "model"}

\item{colNames}{(reactive) use this for colnames of imported data. This parameter is ignored if importType == "model"}

\item{customWarningChecks}{list of reactive(!) functions which will be executed after importing
of data.
 functions need to return TRUE if check is successful or a character with a warning otherwise.
 This parameter is ignored if importType == "model"}

\item{customErrorChecks}{list of reactive(!) functions which will be executed after importing
of data.
  functions need to return TRUE if check is successful or a character with a warning otherwise.
  This parameter is ignored if importType == "model"}

\item{batch}{(logical) use batch import. This parameter is ignored if importType == "model"}

\item{outputAsMatrix}{(logical) TRUE if output must be a matrix,
e.g. for batch = TRUE in Resources. This parameter is ignored if importType == "model"}

\item{fileExtension}{(character) (otional) app specific file extension, e.g. "resources", "bmsc",
"bpred", or (app-unspecific) "zip". Only files with this extension are valid for import.}

\item{mainFolder}{(character) folder containing all loadable .zip files.
This parameter is ignored if importType == "data"}

\item{subFolder}{(character) (optional) subfolder containing loadable .zip files.
This parameter is ignored if importType == "data"}

\item{rPackageName}{(character) If not NULL, than the uploaded file must be a downloaded file
from this R package. This parameter is ignored if importType == "data"}

\item{onlySettings}{(logical) if TRUE allow only upload of user inputs and user data.
This parameter is ignored if importType == "data"}

\item{extractZipFun}{(function) (optional) parameter to provide an app-specific function
to extract the zip object of the upload. This parameter is ignored if importType != "zip".}
}
\description{
Backend for data import module
}
