% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-importData.R
\name{importOptions}
\alias{importOptions}
\title{Import Options}
\usage{
importOptions(rPackageName = "")
}
\arguments{
\item{rPackageName}{(character) If not NULL, than the uploaded file must be a downloaded file
from the R package where \code{importDataServer} is called. This parameter is ignored if
\code{importType == "data"}.}
}
\description{
Extra options for the import module.
}
