% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-importData-selectData.R
\name{loadCKANResourceServer}
\alias{loadCKANResourceServer}
\title{Load CKAN Resource Server}
\usage{
loadCKANResourceServer(id)
}
\arguments{
\item{id}{id of module}
}
\description{
Server function of the module
}
