% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-importData-linkToData.R
\name{loadFileFromLink}
\alias{loadFileFromLink}
\title{Load File From Link}
\usage{
loadFileFromLink(values, loadedSourceInputs, parentParams)
}
\arguments{
\item{values}{(reactiveValues)}

\item{loadedSourceInputs}{(list) user inputs from the dataLink file}

\item{parentParams}{(list) list of parameters from parent module}
}
\description{
Load a file a link points to
}
