% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-loadList.R
\name{loadList}
\alias{loadList}
\title{Load List}
\usage{
loadList(path)
}
\arguments{
\item{path}{(character) path to the file}
}
\value{
(list) list of data for `"json"` or raw data for `"bin"` files
}
\description{
Load a list from a file. Currently supports `"json"` and `"bin"` files.
}
