% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-loadList.R
\name{loadListWrapper}
\alias{loadListWrapper}
\title{Load List Wrapper}
\usage{
loadListWrapper(values, filepath, filename, fileExtension = "json", ...)
}
\arguments{
\item{values}{(list) list with import specifications}

\item{filepath}{(character) url or path}

\item{filename}{(character) name of the model file}

\item{fileExtension}{(character) file type input}

\item{...}{parameters for other wrapper functions}
}
\description{
Wrapper to load a list from a `"json"` file.
}
