% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-importData-loadImport.R
\name{loadModel}
\alias{loadModel}
\title{Load Model}
\usage{
loadModel(filepath, subFolder, rPackageName, onlySettings)
}
\arguments{
\item{filepath}{(character) path to the model file}

\item{subFolder}{(character) (optional) subfolder containing loadable .zip files}

\item{rPackageName}{(character) If not NULL, than the uploaded file must come from this R
package}

\item{onlySettings}{(logical) if TRUE allow only upload of user inputs and user data}
}
\description{
Load Model
}
