% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-importData-loadImport.R
\name{loadModel}
\alias{loadModel}
\title{Load Model}
\usage{
loadModel(filepath, subFolder, rPackageName, onlySettings)
}
\arguments{
\item{filepath}{(character) path to the model file}

\item{subFolder}{(character) name of the (optional) subfolder containing loadable .zip files}

\item{rPackageName}{(character) name of the package (as in the description file) in which this
module is applied, e.g. "mpiBpred"}

\item{onlySettings}{(logical) if TRUE allow only download of user inputs and user data}
}
\description{
Load Model
}
