% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-loadModel.R
\name{loadModelWrapper}
\alias{loadModelWrapper}
\title{Load Model Wrapper}
\usage{
loadModelWrapper(
  filepath,
  subFolder,
  rPackageName,
  onlySettings,
  fileExtension = "zip"
)
}
\arguments{
\item{filepath}{(character) path to the model file}

\item{subFolder}{(character) (optional) subfolder containing loadable .zip files}

\item{rPackageName}{(character) If not NULL, than the uploaded file must come from this R
package}

\item{onlySettings}{(logical) if TRUE allow only upload of user inputs and user data}

\item{fileExtension}{(character) (optional) app specific file extension, e.g. "resources",
"bpred", "bmsc"}
}
\description{
Load Model Wrapper
}
