% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-DataProcessItem-class.R
\name{mapToOldFormat.DataProcessItem}
\alias{mapToOldFormat.DataProcessItem}
\title{S3 method: Map DataProcessItem to old format
Converts a DataProcessItem object to the old format with data and input list.}
\usage{
\method{mapToOldFormat}{DataProcessItem}(obj, ...)
}
\arguments{
\item{obj}{The DataProcessItem object to convert.}
}
\value{
A list with 'data' and 'input' fields, and 'unprocessed' attribute.
}
\description{
S3 method: Map DataProcessItem to old format
Converts a DataProcessItem object to the old format with data and input list.
}
