% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-importData-mergeSettings.R
\name{mergeSettingsUI}
\alias{mergeSettingsUI}
\alias{mergeSettingsServer}
\title{Merge Settings UI}
\usage{
mergeSettingsUI(id)

mergeSettingsServer(id, tableXData, tableYData, tableXId, tableYId)
}
\arguments{
\item{id}{id of module}

\item{tableXData}{(data.frame) data to be merged}

\item{tableYData}{(data.frame) data to be merged}

\item{tableXId}{(character) internal table name, see \link{extractTableIds}}

\item{tableYId}{(character) internal table name, see \link{extractTableIds}}
}
\description{
UI of the merge settings module

Server function of the merge settings module
}
