% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-importData-linkToData.R
\name{observeDownloadDataLink}
\alias{observeDownloadDataLink}
\title{Download a link to import data}
\usage{
observeDownloadDataLink(
  id,
  input,
  output,
  session,
  mergeList,
  downloadBtnID = "downloadDataLink"
)
}
\arguments{
\item{id}{module id}

\item{input}{input object from server function}

\item{output}{output object from server function}

\item{session}{session from server function}

\item{mergeList}{(reactiveVal) list of data imports}

\item{downloadBtnID}{(character) ID of the downloadButton}
}
\description{
Download a link to import data
}
