% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-importData-linkToData.R
\name{observeUploadDataLink}
\alias{observeUploadDataLink}
\title{Download a link to import data}
\usage{
observeUploadDataLink(id, input, output, session)
}
\arguments{
\item{id}{module id}

\item{input}{input object from server function}

\item{output}{output object from server function}

\item{session}{session from server function}
}
\description{
Download a link to import data
}
