% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-importData-linkToData.R
\name{observeUploadDataLink}
\alias{observeUploadDataLink}
\title{Observe Upload of a link to import data}
\usage{
observeUploadDataLink(
  id,
  input,
  output,
  session,
  dataSource,
  parentParams,
  mergeList
)
}
\arguments{
\item{id}{module id}

\item{input}{input object from server function}

\item{output}{output object from server function}

\item{session}{session from server function}

\item{parentParams}{(list) parentParams}

\item{mergeList}{(reactiveVal) list of data imports}
}
\description{
Observe Upload of a link to import data
}
