% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-config.R
\docType{data}
\name{pkg.env}
\alias{pkg.env}
\title{globals}
\format{
An object of class \code{environment} of length 2.
}
\usage{
pkg.env
}
\description{
Contains the package's global variables
}
\section{Fields}{

\describe{
\item{\code{api_key}}{The OpenAI API key}

\item{\code{url.completions}}{The URL for the completions endpoint}
}}

\keyword{datasets}
