% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-importData-prepareData.R
\name{prepareDataServer}
\alias{prepareDataServer}
\title{Prepare Data Server}
\usage{
prepareDataServer(id, mergeList)
}
\arguments{
\item{id}{id of module}

\item{mergeList}{(list) list of selected data}
}
\description{
Server function of the module
}
