% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-importData-previewData.R
\name{previewDataServer}
\alias{previewDataServer}
\title{Preview Data Server}
\usage{
previewDataServer(id, dat)
}
\arguments{
\item{id}{id of module}

\item{dat}{(reactive) data.frame of preview data to be displayed}
}
\description{
Server function of the module
}
