% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-remoteModels.R
\name{remoteModelsServer}
\alias{remoteModelsServer}
\title{Server function for remote models}
\usage{
remoteModelsServer(
  id,
  githubRepo,
  rPackageName,
  rPackageVersion,
  pathToLocalModels = "./predefinedModels"
)
}
\arguments{
\item{id}{namespace id}

\item{githubRepo}{(character) name of used github repository, e.g. "bpred"}

\item{rPackageName}{(character) name of the package (as in the description file) where this
module is applied, e.g. "mpiBpred"}

\item{rPackageVersion}{(character) current version of the package where this module is applied,
e.g. packageVersion("mpiBpred")}

\item{pathToLocalModels}{(character) relative path in the package to the folder storing local
models}
}
\description{
Backend for the module
}
