% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-remoteModels.R
\name{remoteModelsUI}
\alias{remoteModelsUI}
\alias{remoteModelsServer}
\title{Remote models module}
\usage{
remoteModelsUI(
  id,
  selectLabel = "Load online model",
  buttonLabel = "Load",
  width = "100\%"
)

remoteModelsServer(
  id,
  githubRepo,
  pathToLocal = file.path(".", "predefinedModels"),
  folderOnGithub = "/predefinedModels",
  fileExtension = "",
  onlyLocalModels = reactive(FALSE),
  resetSelected = reactive(FALSE),
  reloadChoices = reactive(TRUE),
  rPackageName = NULL,
  rPackageVersion = NULL,
  isInternet = reactive(TRUE)
)
}
\arguments{
\item{id}{id of module}

\item{selectLabel}{label of select input}

\item{buttonLabel}{button label}

\item{width}{width of inputs in percent}

\item{githubRepo}{(character) name of used github repository, e.g. "bpred"}

\item{pathToLocal}{(character) relative path to the folder storing local files}

\item{folderOnGithub}{(character) folder on github where remote files are stored. This should
correspond to 'pathToLocal' since online and offline files should be the same and up-to-date}

\item{fileExtension}{(character) DEPRECATED, not in use anymore. (otional) app specific file
 extension, e.g. "resources", "bmsc",
"bpred", or (app-unspecific) "zip". Only files with this extension are valid for import.}

\item{onlyLocalModels}{(reactive) if TRUE only local files are used instead of accessing github}

\item{resetSelected}{(reactive) if TRUE resets the selected remote file}

\item{reloadChoices}{(reactive) trigger access to  github and reload choices of remote files}

\item{rPackageName}{(character) DEPRECATED (not in use and will be removed in future): name of
the package (as in the description file) in which this module is applied, e.g. "Bpred"}

\item{rPackageVersion}{(character) DEPRECATED (not in use and will be removed in future): current
version of the package where this module is applied, e.g. utils::packageVersion("Bpred")}

\item{isInternet}{(reactive) TRUE if there is an internet connection}
}
\value{
(character) the path to the selected remote (github) or local file
}
\description{
Displays a select input to choose a remote model and a button to load it

Backend for the module
}
