% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-modelUpload-remote.R
\name{remoteModelsServer}
\alias{remoteModelsServer}
\title{Server function for Remote Models}
\usage{
remoteModelsServer(id, githubRepo = "bpred")
}
\arguments{
\item{id}{namespace id}

\item{githubRepo}{(character) name of the app's github repository, currently "bpred",
"osteo-bior", "resources"}
}
\description{
Backend for the remote models module
}
