% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-remoteModels.R
\name{remoteModelsServer}
\alias{remoteModelsServer}
\title{Server function for remote models}
\usage{
remoteModelsServer(
  id,
  githubRepo,
  pathToLocal = file.path(".", "predefinedModels"),
  folderOnGithub = "/predefinedModels",
  onlyLocalModels = reactive(FALSE),
  resetSelected = reactive(FALSE),
  reloadChoices = reactive(FALSE),
  rPackageName = NULL,
  rPackageVersion = NULL
)
}
\arguments{
\item{id}{namespace id}

\item{githubRepo}{(character) name of used github repository, e.g. "bpred"}

\item{pathToLocal}{(character) relative path to the folder storing local models.}

\item{folderOnGithub}{(character) folder on github where remote models are stored. This should
correspond to 'pathToLocal' since online and offline models should be the same and up-to-date}

\item{onlyLocalModels}{(reactive) if TRUE only local models are used}

\item{resetSelected}{(reactive) if TRUE resets the selected remote model}

\item{reloadChoices}{(reactive) trigger access to  github and reload choices of remote models}

\item{rPackageName}{(character) name of the package (as in the description file) in which this
module is applied, e.g. "mpiBpred"}

\item{rPackageVersion}{(character) current version of the package where this module is applied,
e.g. utils::packageVersion("mpiBpred")}
}
\value{
(character) the path to the selected remote (github) or local model
}
\description{
Backend for the module
}
