% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-remoteModels.R
\name{remoteModelsServer}
\alias{remoteModelsServer}
\title{Server function for remote models}
\usage{
remoteModelsServer(
  id,
  githubRepo,
  rPackageName,
  rPackageVersion,
  customPathToLocalModels = NULL,
  onlyLocalModels = reactiveVal(FALSE),
  resetSelected = reactiveVal(FALSE)
)
}
\arguments{
\item{id}{namespace id}

\item{githubRepo}{(character) name of used github repository, e.g. "bpred"}

\item{rPackageName}{(character) name of the package (as in the description file) in which this
module is applied, e.g. "mpiBpred"}

\item{rPackageVersion}{(character) current version of the package where this module is applied,
e.g. utils::packageVersion("mpiBpred")}

\item{customPathToLocalModels}{(character) custom relative path to the folder storing local models.
If assuming that the package 'DataTools' was installed into the same parent directory as the
the app's package in which this module should be applied, than the path for e.g. the package
"mpiBpred" would "../bpred/inst/app/predefinedModels"}

\item{onlyLocalModels}{(reactive) if TRUE only local models are used}

\item{resetSelected}{(reactive) if TRUE resets the selected remote model.}
}
\value{
(character) the path to the selected remote (github) or local model
}
\description{
Backend for the module
}
