% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-remoteModels.R
\name{remoteModelsServer}
\alias{remoteModelsServer}
\title{Server function for remote models}
\usage{
remoteModelsServer(
  id,
  githubRepo,
  pathToLocal = file.path(".", "predefinedModels"),
  folderOnGithub = "/predefinedModels",
  fileExtension = "zip",
  onlyLocalModels = reactive(FALSE),
  resetSelected = reactive(FALSE),
  reloadChoices = reactive(TRUE),
  rPackageName = NULL,
  rPackageVersion = NULL,
  isInternet = reactive(TRUE)
)
}
\arguments{
\item{id}{namespace id}

\item{githubRepo}{(character) name of used github repository, e.g. "bpred"}

\item{pathToLocal}{(character) relative path to the folder storing local files}

\item{folderOnGithub}{(character) folder on github where remote files are stored. This should
correspond to 'pathToLocal' since online and offline files should be the same and up-to-date}

\item{fileExtension}{(character) (otional) app specific file extension, e.g. "resources", "bmsc",
"bpred", or (app-unspecific) "zip". Only files with this extension are valid for import.}

\item{onlyLocalModels}{(reactive) if TRUE only local files are used instead of accessing github}

\item{resetSelected}{(reactive) if TRUE resets the selected remote file}

\item{reloadChoices}{(reactive) trigger access to  github and reload choices of remote files}

\item{rPackageName}{(character) DEPRECATED (not in use and will be removed in future): name of
the package (as in the description file) in which this module is applied, e.g. "mpiBpred"}

\item{rPackageVersion}{(character) DEPRECATED (not in use and will be removed in future): current
version of the package where this module is applied, e.g. utils::packageVersion("mpiBpred")}

\item{isInternet}{(reactive) TRUE if there is an internet connection}
}
\value{
(character) the path to the selected remote (github) or local file
}
\description{
Backend for the module
}
