% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-modelUpload-remote.R
\name{remoteModelsUI}
\alias{remoteModelsUI}
\title{Remote Models module}
\usage{
remoteModelsUI(id)
}
\arguments{
\item{id}{id of module}
}
\description{
Displays a button which opens a import dialog when clicked
}
