% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-remoteModels.R
\name{remoteModelsUI}
\alias{remoteModelsUI}
\title{Remote models module}
\usage{
remoteModelsUI(
  id,
  selectLabel = "Load online model",
  buttonLabel = "Load",
  width = "100\%"
)
}
\arguments{
\item{id}{id of module}

\item{selectLabel}{label of select input}

\item{buttonLabel}{button label}

\item{width}{width of inputs in percent}
}
\description{
Displays a select input to choose a remote model and a button to load it
}
