% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-importData-selectData.R
\name{selectDataServer}
\alias{selectDataServer}
\title{Select Data Server}
\usage{
selectDataServer(
  id,
  mergeList,
  rowNames = reactiveVal(NULL),
  colNames = reactiveVal(NULL),
  ignoreWarnings = FALSE
)
}
\arguments{
\item{id}{id of module}

\item{mergeList}{(list) list of (previously) selected data}

\item{rowNames}{(reactive) use this for rownames of imported data}

\item{colNames}{(reactive) use this for colnames of imported data}

\item{ignoreWarnings}{TRUE to enable imports in case of warnings}
}
\description{
Server function of the module
}
