% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-importData-selectData.R
\name{selectDataServer}
\alias{selectDataServer}
\title{Select Data Server}
\usage{
selectDataServer(
  id,
  mergeList,
  customNames,
  openPopupReset,
  internetCon,
  ignoreWarnings = FALSE
)
}
\arguments{
\item{id}{id of module}

\item{mergeList}{(list) list of selected data}

\item{customNames}{settings for custom column and row names}

\item{openPopupReset}{(reactive) if TRUE reset ckan source inputs}

\item{ignoreWarnings}{TRUE to enable imports in case of warnings}
}
\description{
Server function of the module
}
