% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-importData-selectData.R
\name{selectDataServer}
\alias{selectDataServer}
\title{Select Data Server}
\usage{
selectDataServer(id, customNames)
}
\arguments{
\item{id}{id of module}

\item{customNames}{customNames}
}
\description{
Server function of the module
}
