% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-importData-selectData.R
\name{selectDataServer}
\alias{selectDataServer}
\title{Select Data Server}
\usage{
selectDataServer(
  id,
  importType,
  mergeList,
  customNames,
  openPopupReset,
  internetCon,
  ckanFileTypes = c("xls", "xlsx", "csv", "odt", "txt"),
  mainFolder = "predefinedModels",
  subFolder = NULL,
  ignoreWarnings = FALSE,
  rPackageName = "",
  onlySettings = FALSE,
  fileExtension = "zip",
  expectedFileInZip = c()
)
}
\arguments{
\item{id}{id of module}

\item{importType}{(character) type of import, either "data" or "model" or "zip".
ImportType == "zip" enables the optional parameter 'expectedFileInZip'.}

\item{mergeList}{(list) list of selected data}

\item{customNames}{settings for custom column and row names}

\item{openPopupReset}{(reactive) if TRUE reset ckan source inputs}

\item{internetCon}{(reactive) TRUE if there is an internet connection}

\item{ckanFileTypes}{(character) file types allowed for import from Pandora ("ckan")}

\item{mainFolder}{(character) folder containing all loadable .zip files.
This parameter is ignored if importType == "data"}

\item{subFolder}{(character) (optional) subfolder containing loadable .zip files.
This parameter is ignored if importType == "data"}

\item{ignoreWarnings}{TRUE to enable imports in case of warnings}

\item{rPackageName}{(character) DEPRECATED. Instead, please use
\code{options = importOptions(rPackageName = <your package>)}.}

\item{onlySettings}{(logical) if TRUE allow only upload of user inputs and user data.
This parameter is ignored if importType == "data"}

\item{fileExtension}{(character) (otional) app specific file extension, e.g. "resources", "bmsc",
"bpred", or (app-unspecific) "zip". Only files with this extension are valid for import.}

\item{expectedFileInZip}{(character) (optional) This parameter is ignored if importType != "zip".
File names that must be contained in the zip upload.}
}
\description{
Server function of the module
}
