% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-importData-selectData.R
\name{selectDataServer}
\alias{selectDataServer}
\title{Select Data Server}
\usage{
selectDataServer(
  id,
  mergeList,
  customNames,
  openPopupReset,
  internetCon,
  githubRepo,
  mainFolder = "predefinedModels",
  subFolder = NULL,
  ignoreWarnings = FALSE,
  rPackageName = NULL,
  onlySettings = FALSE
)
}
\arguments{
\item{id}{id of module}

\item{mergeList}{(list) list of selected data}

\item{customNames}{settings for custom column and row names}

\item{openPopupReset}{(reactive) if TRUE reset ckan source inputs}

\item{internetCon}{(reactive) TRUE if there is an internet connection}

\item{githubRepo}{(character) name of used github repository, e.g. "bpred"}

\item{mainFolder}{(character) folder containing all loadable .zip files}

\item{subFolder}{(character) (optional) subfolder containing loadable .zip files}

\item{ignoreWarnings}{TRUE to enable imports in case of warnings}

\item{rPackageName}{(character) If not NULL, than the uploaded file must come from this R
package}

\item{onlySettings}{(logical) if TRUE allow only upload of user inputs and user data}
}
\description{
Server function of the module
}
