% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-importData.R
\name{selectDataTab}
\alias{selectDataTab}
\title{Select Data UI}
\usage{
selectDataTab(ns, defaultSource = "ckan", batch = FALSE)
}
\arguments{
\item{ns}{namespace}

\item{defaultSource}{(character) default source for input "Source", e.g. "ckan", "file", or "url"}

\item{batch}{(logical) use batch import}
}
\description{
Select Data UI
}
