% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-importData-selectData.R
\name{selectDataUI}
\alias{selectDataUI}
\title{Select Data UI}
\usage{
selectDataUI(id, batch, outputAsMatrix, importType, options = importOptions())
}
\arguments{
\item{id}{id of module}

\item{batch}{(logical) use batch import. This parameter is ignored if importType == "model"}

\item{outputAsMatrix}{(logical) TRUE if output must be a matrix,
e.g. for batch = TRUE in Resources. This parameter is ignored if importType == "model"}

\item{importType}{(character) type of import, either "data" or "model" or "zip".
ImportType == "zip" enables the optional parameter 'expectedFileInZip'.}

\item{options}{(list) Extra options for the import module.}

\item{isInternet}{(logical) set TRUE, if there is an internet connection. This parameter is
ignored if \code{type = "file"} or \code{type = "remoteModel"}}
}
\description{
UI of the module
}
