% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-importData-selectData.R
\name{selectFileTypeUI}
\alias{selectFileTypeUI}
\title{Select File Type UI}
\usage{
selectFileTypeUI(id, importType)
}
\arguments{
\item{id}{id of module}

\item{importType}{(character) type of import, either "data" or "model"}
}
\description{
UI of the module
}
