% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-importData-selectData.R
\name{selectSourceServer}
\alias{selectSourceServer}
\title{Select Source Server}
\usage{
selectSourceServer(id, openPopupReset, defaultSource)
}
\arguments{
\item{id}{id of module}

\item{openPopupReset}{(reactive) if TRUE reset ckan source inputs}

\item{defaultSource}{(character) default source for input "Source", e.g. "ckan", "file", or "url"}
}
\description{
Server function of the module
}
