% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-importData-selectData.R
\name{selectSourceServer}
\alias{selectSourceServer}
\title{Select Source Server}
\usage{
selectSourceServer(
  id,
  openPopupReset,
  internetCon,
  githubRepo,
  mainFolder,
  subFolder
)
}
\arguments{
\item{id}{id of module}

\item{openPopupReset}{(reactive) if TRUE reset ckan source inputs}

\item{internetCon}{(reactive) TRUE if there is an internet connection}

\item{githubRepo}{(character) name of used github repository, e.g. "bpred"}

\item{mainFolder}{(character) folder containing all loadable .zip files}

\item{subFolder}{(character) (optional) subfolder containing loadable .zip files}
}
\description{
Server function of the module
}
