% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-importData-selectData.R
\name{selectSourceServer}
\alias{selectSourceServer}
\title{Select Source Server}
\usage{
selectSourceServer(
  id,
  openPopupReset,
  internetCon,
  githubRepo,
  ckanFileTypes = c("xls", "xlsx", "csv", "odt", "txt"),
  exampleOptions = exampleOptions()
)
}
\arguments{
\item{id}{id of module}

\item{openPopupReset}{(reactive) if TRUE reset ckan source inputs}

\item{internetCon}{(reactive) TRUE if there is an internet connection}

\item{githubRepo}{(character) name of used github repository, e.g. "bpred"}

\item{ckanFileTypes}{(character) file types allowed for import from Pandora ("ckan")}

\item{exampleOptions}{extra options for remote or local examples}
}
\description{
Server function of the module
}
