% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-importData-selectData.R
\name{selectSourceServer}
\alias{selectSourceServer}
\title{Select Source Server}
\usage{
selectSourceServer(
  id,
  openPopupReset,
  internetCon,
  githubRepo,
  folderOnGithub,
  pathToLocal,
  ckanFileTypes = c("xls", "xlsx", "csv", "odt", "txt")
)
}
\arguments{
\item{id}{id of module}

\item{openPopupReset}{(reactive) if TRUE reset ckan source inputs}

\item{internetCon}{(reactive) TRUE if there is an internet connection}

\item{githubRepo}{(character) name of used github repository, e.g. "bpred"}

\item{folderOnGithub}{(character) folder on github where remote files are stored. This should
correspond to 'pathToLocal' since online and offline files should be the same and up-to-date}

\item{pathToLocal}{(character) relative path to the folder storing local files}

\item{ckanFileTypes}{(character) file types allowed for import from Pandora ("ckan")}
}
\description{
Server function of the module
}
