% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-importData-selectData.R
\name{selectSourceUI}
\alias{selectSourceUI}
\title{Select Source UI}
\usage{
selectSourceUI(id, defaultSource, sourceChoices)
}
\arguments{
\item{id}{id of module}

\item{defaultSource}{(character) default source for input "Source", e.g. "ckan", "file", or "url"}

\item{sourceChoices}{(character) named list of choices for the input 'Source'}
}
\description{
UI of the module
}
