% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-importData-selectSource.R
\name{selectSourceUI}
\alias{selectSourceUI}
\title{Select Source UI}
\usage{
selectSourceUI(
  id,
  defaultSource,
  ckanFileTypes,
  importType,
  isInternet,
  fileExtension = "zip"
)
}
\arguments{
\item{id}{id of module}

\item{defaultSource}{(character) default source for input "Source", e.g. "ckan", "file", or "url"}

\item{ckanFileTypes}{(character) file types allowed for import from Pandora ("ckan")}

\item{importType}{(character) type of import, either "data" or "model" or "zip".
ImportType == "zip" enables the optional parameter 'expectedFileInZip'.}

\item{isInternet}{(logical) set TRUE, if there is an internet connection. This parameter is
ignored if \code{type = "file"} or \code{type = "remoteModel"}}

\item{fileExtension}{(character) (otional) app specific file extension, e.g. "resources", "bmsc",
"bpred", or (app-unspecific) "zip". Only files with this extension are valid for import.}
}
\description{
UI of the module
}
