% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-toolsPanel.R
\name{toolsImportUI}
\alias{toolsImportUI}
\alias{toolsImportServer}
\title{UI function of toolsImport module}
\usage{
toolsImportUI(id)

toolsImportServer(id)
}
\arguments{
\item{id}{module id}
}
\description{
Module for DataTool's example app
}
