% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-toolsPanel.R
\name{toolsImportServer}
\alias{toolsImportServer}
\title{Server function of toolsImport module}
\usage{
toolsImportServer(id, config)
}
\arguments{
\item{id}{module id}

\item{config}{(list) list of config variables}
}
\description{
Server function of toolsImport module
}
