% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-toolsPanel.R
\name{toolsPanelServer}
\alias{toolsPanelServer}
\title{Server function of toolsPanel module}
\usage{
toolsPanelServer(id, defaultSource = "ckan")
}
\arguments{
\item{id}{module id}

\item{defaultSource}{(character) default source for input "Source", e.g. "ckan", "file", or "url"}
}
\description{
Server function of toolsPanel module
}
