% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-tryCatchWithErrorsAndWarnings.R
\name{tryCatchWithWarningsAndErrors}
\alias{tryCatchWithWarningsAndErrors}
\title{Try Catch with Warnings and Errors}
\usage{
tryCatchWithWarningsAndErrors(expr, messagePreError = "Modeling failed:")
}
\arguments{
\item{expr}{expression to be evaluated.}

\item{messagePreError}{(character) error message prefix.}
}
\description{
Catches errors and warnings and forwards them as an alert. If an error occurs, NULL is returned.
If a warning occurs, the result is returned.
}
