% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-DataProcessItem-class.R
\name{update.DataProcessItem}
\alias{update.DataProcessItem}
\title{S3 method: Update fields of a DataProcessItem
Updates fields of a DataProcessItem object.}
\usage{
\method{update}{DataProcessItem}(
  item,
  data = NULL,
  input = NULL,
  unprocessed = NULL,
  filename = NULL,
  history = NULL,
  ...
)
}
\arguments{
\item{item}{The DataProcessItem object to update.}

\item{data}{New data to set (optional).}

\item{input}{New input list to set (optional).}

\item{unprocessed}{New unprocessed status to set (optional).}

\item{filename}{New filename to set (optional).}

\item{history}{New history to set (optional).}
}
\value{
The updated DataProcessItem object.
}
\description{
S3 method: Update fields of a DataProcessItem
Updates fields of a DataProcessItem object.
}
