% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-importData-linkToData.R
\name{updateUserInputs}
\alias{updateUserInputs}
\title{Update User Inputs}
\usage{
updateUserInputs(id, input, output, session, userInputs)
}
\arguments{
\item{id}{module id}

\item{input}{input object from server function}

\item{output}{output object from server function}

\item{session}{session from server function}

\item{userInputs}{(list) list of inputs to be updated}
}
\description{
Update User Inputs
}
