% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-downAndUpload.R
\name{uploadModelServer}
\alias{uploadModelServer}
\title{Server function upload model}
\usage{
uploadModelServer(
  id,
  githubRepo,
  mainFolder = "predefinedModels",
  subFolder = NULL,
  rPackageName = NULL,
  reloadChoices = reactive(FALSE),
  onlySettings = FALSE,
  reset = reactive(FALSE)
)
}
\arguments{
\item{id}{namespace id}

\item{githubRepo}{(character) name of used github repository, e.g. "bpred"}

\item{mainFolder}{(character) folder containing all loadable .zip files}

\item{subFolder}{(character) (optional) subfolder containing loadable .zip files}

\item{rPackageName}{(character) If not NULL, than the uploaded file must come from this R
package}

\item{reloadChoices}{(reactive) trigger access to  github and reload choices of remote files}

\item{onlySettings}{(logical) if TRUE allow only download of user inputs and user data}

\item{reset}{(reactive) resets the selection of the online files}
}
\description{
Backend for upload model module
}
