% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-downAndUpload.R
\name{uploadModelServer}
\alias{uploadModelServer}
\title{Server function upload model}
\usage{
uploadModelServer(
  id,
  githubRepo,
  rPackageName,
  onlySettings,
  reset = reactive(FALSE)
)
}
\arguments{
\item{id}{namespace id}

\item{githubRepo}{(character) name of used github repository, e.g. "bpred"}

\item{rPackageName}{(character) name of the package (as in the description file) in which this
module is applied, e.g. "mpiBpred"}

\item{onlySettings}{(logical) if TRUE allow only download of user inputs and user data}

\item{reset}{(reactive) resets the selection of the online model}
}
\description{
Backend for upload model module
}
