% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-downAndUpload.R
\name{uploadModelServer}
\alias{uploadModelServer}
\title{Server function upload model}
\usage{
uploadModelServer(
  id,
  githubRepo,
  modelFolder = "predefinedModels",
  modelSubFolder = NULL,
  reloadChoices = reactive(FALSE),
  onlySettings = FALSE,
  reset = reactive(FALSE)
)
}
\arguments{
\item{id}{namespace id}

\item{githubRepo}{(character) name of used github repository, e.g. "bpred"}

\item{modelFolder}{(character) folder containing all predefined models}

\item{reloadChoices}{(reactive) trigger access to  github and reload choices of remote models}

\item{onlySettings}{(logical) if TRUE allow only download of user inputs and user data}

\item{reset}{(reactive) resets the selection of the online model}

\item{subFolder}{(character) possible subfolder containing predefined models}
}
\description{
Backend for upload model module
}
