% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-downAndUpload.R
\name{uploadModelServer}
\alias{uploadModelServer}
\title{Server function upload model}
\usage{
uploadModelServer(
  id,
  githubRepo,
  mainFolder = "predefinedModels",
  subFolder = NULL,
  rPackageName = NULL,
  reloadChoices = reactive(TRUE),
  onlySettings = FALSE,
  fileExtension = "zip",
  reset = reactive(FALSE)
)
}
\arguments{
\item{id}{namespace id}

\item{githubRepo}{(character) name of used github repository, e.g. "bpred"}

\item{mainFolder}{(character) folder containing all loadable .zip files. For most apps this
is folder "predefinedModels". In most apps it can be found under "inst/app/".}

\item{subFolder}{(character) (optional) subfolder containing loadable .zip files}

\item{rPackageName}{(character) If not NULL, than the uploaded file must come from this R
package}

\item{reloadChoices}{(reactive) trigger access to  github and reload choices of remote files}

\item{onlySettings}{(logical) if TRUE allow only upload of user inputs and user data}

\item{fileExtension}{(character) (optional) app specific file extension, e.g. "resources",
"bpred", "bmsc"}

\item{reset}{(reactive) resets the selection of the online files}
}
\description{
Backend for upload model module
}
