% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-downAndUpload.R
\name{uploadModelServer}
\alias{uploadModelServer}
\title{Server function upload model}
\usage{
uploadModelServer(
  id,
  githubRepo,
  folderOnGithub = "/predefinedModels",
  pathToLocal = file.path(".", "predefinedModels"),
  reloadChoices = reactive(FALSE),
  onlySettings = FALSE,
  reset = reactive(FALSE)
)
}
\arguments{
\item{id}{namespace id}

\item{githubRepo}{(character) name of used github repository, e.g. "bpred"}

\item{folderOnGithub}{(character) folder on github where remote models are stored. This should
correspond to 'pathToLocal' since online and offline models should be the same and up-to-date}

\item{pathToLocal}{(character) relative path to the folder storing local models.}

\item{reloadChoices}{(reactive) trigger access to  github and reload choices of remote models}

\item{onlySettings}{(logical) if TRUE allow only download of user inputs and user data}

\item{reset}{(reactive) resets the selection of the online model}
}
\description{
Backend for upload model module
}
