% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-downAndUpload.R
\name{uploadModelUI}
\alias{uploadModelUI}
\title{Upload model module}
\usage{
uploadModelUI(
  id,
  label,
  labelRemote = "Load online model",
  labelLocal = "Load local model",
  width = NULL
)
}
\arguments{
\item{id}{id of module}

\item{label}{title of module}

\item{labelRemote}{label of the input for remote files}

\item{labelLocal}{label of the input for local files}

\item{width}{width of inputs in percent}
}
\description{
UI function to upload a zip file with exportNotes and a list of models
}
