% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-modelUpload.R
\name{uploadModelUI}
\alias{uploadModelUI}
\title{Upload model module}
\usage{
uploadModelUI(id, label = "Upload Model")
}
\arguments{
\item{id}{id of module}

\item{label}{label of module}
}
\description{
UI function to upload a zip file with notes and a list of models. Uploads are possible either
from a local source, or from a selection of remote models (stored on github or locally).
}
